### 2024-08-08 #7089 @editorialbot
[PRE REVIEW]: reflectorch: a deep learning package for X-ray and neutron reflectometry
**Submitting author:** @valentinsingularity (<a href="http://orcid.org/0000-0002-5897-3863">Valentin Munteanu</a>)
**Repository:** https://github.com/schreiber-lab/reflectorch
**Branch with paper.md** (empty if default branch):
**Version:** v1.2.1
