+++ configs/b_mc_point_neutron_conv_standard_L2_comp.yaml
@@ -1,0 +1,106 @@
+general:
+  name: b_mc_point_neutron_conv_standard_L2_comp
+  root_dir: null
+  
+dset:
+  cls: ReflectivityDataLoader
+  prior_sampler:
+    cls: SubpriorParametricSampler
+    kwargs:
+      param_ranges:
+        thicknesses: [1., 500.]
+        roughnesses: [0., 60.]
+        slds: [-8., 16.]
+      bound_width_ranges:
+        thicknesses: [1.0e-2, 500.]
+        roughnesses: [1.0e-2, 60.]
+        slds: [1.0e-2, 5.]
+      model_name: standard_model
+      max_num_layers: 2
+      constrained_roughness: true
+      max_thickness_share: 0.5
+      logdist: false
+      scale_params_by_ranges: false
+      scaled_range: [-1., 1.]
+      device: 'cuda'
+      
+  q_generator:
+    cls: ConstantQ
+    kwargs:
+      q: [0.005, 0.2, 128]
+      device: 'cuda'
+
+  intensity_noise:
+    cls: BasicExpIntensityNoise
+    kwargs:
+      relative_errors: [0.0, 0.2]
+      abs_errors: 0.0
+      consistent_rel_err: false
+      logdist: false
+      apply_shift: true
+      shift_range: [-0.3, 0.3]
+      apply_scaling: true
+      scale_range: [-0.02, 0.02]
+      apply_background: true
+      background_range: [1.0e-10, 1.0e-4]
+      add_to_context: true
+
+  smearing:
+    cls: Smearing
+    kwargs:
+      sigma_range: [0.01, 0.10]
+      constant_dq: False
+      gauss_num: 17
+      share_smeared: 0.8
+      
+  curves_scaler:
+    cls: LogAffineCurvesScaler
+    kwargs:
+      weight: 0.2
+      bias: 1.0
+      eps: 1.0e-10
+
+model:
+  network:
+    cls: NetworkWithPriorsConvEmb
+    pretrained_name: null
+    device: 'cuda'
+    kwargs:
+      in_channels: 1
+      hidden_channels: [32, 64, 128, 256, 512]
+      dim_embedding: 128
+      dim_avpool: 1
+      embedding_net_activation: 'gelu'
+      use_batch_norm: true
+      dim_out: 8
+      layer_width: 512
+      num_blocks: 8
+      repeats_per_block: 2
+      mlp_activation: 'gelu'
+      dropout_rate: 0.0
+      conditioning: 'film' 
+      pretrained_embedding_net: null
+       
+training:
+  num_iterations: 100000
+  batch_size: 4096
+  lr: 1.0e-3
+  grad_accumulation_steps: 1
+  clip_grad_norm_max: null
+  update_tqdm_freq: 1
+  optimizer: AdamW
+  trainer_kwargs:
+    optim_kwargs:
+      betas: [0.9, 0.999]
+      weight_decay: 0.0005
+  callbacks:
+    save_best_model:
+      enable: true
+      freq: 500
+    lr_scheduler:
+      cls: StepLR
+      kwargs:
+        step_size: 50000
+        gamma: 0.1
+  logger:
+    use_neptune: false
