+++ configs/a_base_point_xray_conv_standard.yaml
@@ -1,0 +1,133 @@
+general:
+  name: a_base_point_xray_conv_standard
+  root_dir: null
+  
+dset:
+  cls: ReflectivityDataLoader
+  prior_sampler:
+    cls: SubpriorParametricSampler
+    kwargs:
+      param_ranges:
+        thicknesses: [1., 500.]
+        roughnesses: [0., 60.]
+        slds: [0., 50.]
+      bound_width_ranges:
+        thicknesses: [1.0e-2, 500.]
+        roughnesses: [1.0e-2, 60.]
+        slds: [1.0e-2, 5.]
+      model_name: standard_model
+      max_num_layers: 2
+      constrained_roughness: true
+      max_thickness_share: 0.5
+      logdist: false
+      scale_params_by_ranges: false
+      scaled_range: [-1., 1.]
+      device: 'cuda'
+      
+  q_generator:
+    cls: ConstantQ
+    kwargs:
+      q: [0.02, 0.15, 128]
+      device: 'cuda'
+
+  # q_generator:
+  #   cls: VariableQ
+  #   kwargs:
+  #     q_min_range: [0.005, 0.03]
+  #     q_max_range: [0.15, 0.3]
+  #     n_q_range: [128, 128]
+  #     device: 'cuda'
+      
+  # intensity_noise:
+  #   cls: GaussianExpIntensityNoise
+  #   kwargs:
+  #     relative_errors: [0.01, 0.3]
+  #     consistent_rel_err: false
+  #     apply_shift: true
+  #     shift_range: [-0.3, 0.3]
+  #     add_to_context: true
+
+  intensity_noise:
+    cls: BasicExpIntensityNoise
+    kwargs:
+      relative_errors: [0.0, 0.2]
+      abs_errors: 0.0
+      consistent_rel_err: false
+      logdist: false
+      apply_shift: true
+      shift_range: [-0.3, 0.3]
+      apply_scaling: false
+      scale_range: [-0.02, 0.02]
+      add_to_context: true
+ 
+  curves_scaler:
+    cls: LogAffineCurvesScaler
+    kwargs:
+      weight: 0.2
+      bias: 1.0
+      eps: 1.0e-10
+
+model:
+  network:
+    cls: NetworkWithPriors
+    pretrained_name: null
+    device: 'cuda'
+    kwargs:
+      embedding_net_type: 'conv'
+      embedding_net_kwargs:
+        in_channels: 1
+        hidden_channels: [32, 64, 128, 256, 512]
+        kernel_size: 3
+        dim_embedding: 128
+        dim_avpool: 1
+        use_batch_norm: true
+        use_se: false
+        activation: 'gelu'
+      pretrained_embedding_net: null
+      dim_out: 8
+      dim_conditioning_params: 0
+      layer_width: 512
+      num_blocks: 8
+      repeats_per_block: 2
+      residual: true
+      use_batch_norm: true
+      use_layer_norm: false
+      mlp_activation: 'gelu'
+      dropout_rate: 0.0 
+      tanh_output: false
+      conditioning: 'film' # 'concat', 'glu', 'film'
+      concat_condition_first_layer: false
+       
+training:
+  trainer_cls: PointEstimatorTrainer
+  num_iterations: 100000
+  batch_size: 4096
+  lr: 1.0e-3
+  grad_accumulation_steps: 1
+  clip_grad_norm_max: null
+  update_tqdm_freq: 1
+  optimizer: AdamW
+  trainer_kwargs:
+    train_with_q_input: false
+    train_with_sigmas: false
+    condition_on_q_resolutions: false
+    rescale_loss_interval_width: true
+    use_l1_loss: true
[…]
