### 2025-03-06 6736d8e
added GaussianExpIntensityNoise
