+++ reflectorch/data_generation/noise.py
@@ -1,11 +1,11 @@
-from typing import Union, Tuple
+from typing import List, Union, Tuple
 from math import log10
 
 import torch
 from torch import Tensor
 
 from reflectorch.data_generation.process_data import ProcessData
-from reflectorch.data_generation.utils import uniform_sampler
+from reflectorch.data_generation.utils import logdist_sampler, uniform_sampler
 
 __all__ = [
     "QNoiseGenerator",
@@ -18,6 +18,7 @@ __all__ = [
     "ShiftNoise",
     "BackgroundNoise",
     "BasicExpIntensityNoise",
+    "GaussianExpIntensityNoise",
     "BasicQNoiseGenerator",
 ]
 
@@ -102,18 +103,22 @@ class BasicQNoiseGenerator(QNoiseGenerator):
                                     Defaults to (0, 1e-3).
     """
     def __init__(self,
+                 apply_systematic_shifts: bool = True,
                  shift_std: float = 1e-3,
+                 apply_gaussian_noise: bool = False,
                  noise_std: Union[float, Tuple[float, float]] = (0, 1e-3),
                  add_to_context: bool = False,
                  ):
-        self.q_shift = QSystematicShiftGenerator(shift_std, add_to_context=add_to_context)
-        self.q_noise = QNormalNoiseGenerator(noise_std, add_to_context=add_to_context)
+        self.q_shift = QSystematicShiftGenerator(shift_std, add_to_context=add_to_context) if apply_systematic_shifts else None
+        self.q_noise = QNormalNoiseGenerator(noise_std, add_to_context=add_to_context) if apply_gaussian_noise else None
 
     def apply(self, qs: Tensor, context: dict = None):
-        """applies random noise to the q values"""
+        """applies noise to the q values"""
         qs = torch.atleast_2d(qs)
-        qs = self.q_shift.apply(qs, context)
-        qs = self.q_noise.apply(qs, context)
+        if self.q_shift:
+            qs = self.q_shift.apply(qs, context)
+        if self.q_noise:
+            qs = self.q_noise.apply(qs, context)
         return qs
 
 
@@ -154,6 +159,51 @@ class MultiplicativeLogNormalNoiseGenerator(IntensityNoiseGenerator):
 
         return noise * curves
 
+class GaussianNoiseGenerator(IntensityNoiseGenerator):
+    """Noise generator which applies noise as R_n = R + eps, with eps~N(0, sigmas) and sigmas = relative_errors * R
+
+    Args:
+        relative_errors (Union[float, Tuple[float, float], List[Tuple[float, float]]])
+        consistent_relative_errors (bool): If True the relative_error is the same for all point of a curve, otherwise it is sampled uniformly.
+    """
+
+    def __init__(self, relative_errors: Union[float, Tuple[float, float], List[float], List[Tuple[float, float]]], 
+                 consistent_rel_err: bool = False,
+                 add_to_context: bool = False):
+        self.relative_errors = relative_errors
+        self.consistent_rel_err = consistent_rel_err
+        self.add_to_context = add_to_context
+
+    def apply(self, curves: Tensor, context: dict = None):
+        """Applies Gaussian noise to the curves."""
+        relative_errors = self.relative_errors
+        num_channels = curves.shape[1] if curves.dim() == 3 else 1
+
+        if isinstance(relative_errors, float):
+            relative_errors = torch.ones_like(curves) * relative_errors
+
+        elif isinstance(relative_errors, (list, tuple)) and isinstance(relative_errors[0], float):
+            if self.consistent_rel_err:
+                relative_errors = uniform_sampler(*relative_errors, curves.shape[0], num_channels, device=curves.device, dtype=curves.dtype)
+                if num_channels > 1:
+                    relative_errors = relative_errors.unsqueeze(-1)
+            else:
[…]
