+++ reflectorch/ml/__init__.py
@@ -15,6 +15,7 @@ __all__ = [
     'Logger',
     'Loggers',
     'PrintLogger',
+    'TensorBoardLogger',
     'ScheduleBatchSize',
     'ScheduleLR',
     'StepLR',
@@ -22,6 +23,7 @@ __all__ = [
     'LogCyclicLR',
     'ReduceLROnPlateau',
     'OneCycleLR',
+    'CosineAnnealingWithWarmup',
     'ReflectivityDataLoader',
     'MultilayerDataLoader',
     'RealTimeSimTrainer',
