### 2025-03-06 45eee6f
modified q generators
