### 2025-03-06 aff478a
fixed tqdm bar; removed train_with_q_input argument
