### 2025-03-06 be865fc
extended LogLosses for logging multiple losses
