### 2025-03-06 969ecf2
added CosineAnnealingWithWarmup scheduler; added plotting functionality for schedulers
