+++ reflectorch/ml/schedulers.py
@@ -12,6 +12,7 @@ __all__ = [
     'LogCyclicLR',
     'ReduceLROnPlateau',
     'OneCycleLR',
+    'CosineAnnealingWithWarmup',
 ]
 
 
@@ -68,6 +69,31 @@ class ScheduleLR(TrainerCallback):
             batch_num (int): index of the current iteration
         """
         self.lr_scheduler.step()
+
+    def simulate_and_plot(self, total_steps: int, initial_lr: float, log_scale: bool = False):
+        import torch
+        import matplotlib.pyplot as plt
+
+        dummy_optim = torch.optim.Adam([torch.zeros(1)], lr=initial_lr)
+        scheduler = self.lr_scheduler_cls(dummy_optim, **self.kwargs)
+
+        lrs = []
+        for step in range(total_steps):
+            lrs.append(dummy_optim.param_groups[0]['lr'])
+            scheduler.step()
+
+        plt.figure(figsize=(10, 6))
+        plt.plot(lrs, label='Learning Rate')
+        plt.xlabel('Steps')
+        plt.ylabel('Learning Rate')
+        plt.title('Learning Rate Schedule')
+
+        if log_scale:
+            plt.yscale('log')
+
+        plt.grid(True, which="both", linestyle='--', linewidth=0.5)
+        plt.legend()
+        plt.show()
 
 
 class StepLR(ScheduleLR):
@@ -176,6 +202,21 @@ class LogCyclicLR(TrainerCallback):
         for param_group in self.param_groups:
             trainer.set_lr(lr, param_group)
 
+    def simulate_and_plot(self, total_steps: int, log_scale: bool = True):
+        import matplotlib.pyplot as plt
+        lrs = [self.get_lr(batch_num) for batch_num in range(total_steps)]
+
+        plt.figure(figsize=(10, 6))
+        plt.plot(lrs, label='Learning Rate')
+        plt.xlabel('Steps')
+        plt.ylabel('Learning Rate')
+        plt.title('Learning Rate Schedule')
+        if log_scale:
+            plt.yscale('log')
+        plt.grid(True, which='both', linestyle='--', linewidth=0.5)
+        plt.legend()
+        plt.show()
+
 
 class ReduceLROnPlateau(TrainerCallback):
     """Learning rate scheduler which reduces the learning rate when the loss stops decreasing
@@ -238,3 +279,77 @@ class OneCycleLR(ScheduleLR):
             three_phase=three_phase,
             **kwargs
         )
+
+class CosineAnnealingWithWarmup(TrainerCallback):
+    """
+    Cosine annealing scheduler with a warm-up stage.
+
+    Args:
+        max_lr (float): The maximum learning rate after the warm-up phase.
+        min_lr (float): The minimum learning rate after the warm-up phase.
+        warmup_iters (int): The number of iterations for the warm-up phase.
+        total_iters (int): The total number of iterations for the scheduler (including warm-up).
+    """
+    def __init__(self, max_lr=None, min_lr=1.0e-6, warmup_iters=100, total_iters=100000):
+        self.max_lr = max_lr
+        self.min_lr = min_lr
+        self.warmup_iters = warmup_iters
+        self.total_iters = total_iters
+
+    def get_lr(self, step):
+        """
+        Compute the learning rate for a given iteration.
+
+        Args:
+            step (int): The current iteration.
+
+        Returns:
+            float: The learning rate for the current iteration.
+        """
+        if step < self.warmup_iters:
+            # Warm-up stage: Linear increase from 0 to max_lr
+            return self.max_lr * step / self.warmup_iters
+        elif step < self.total_iters:
+            # Cosine annealing stage
+            t = (step - self.warmup_iters) / (self.total_iters - self.warmup_iters)
[…]
