### 2025-03-06 260cf89
modified functionality for logging; removed train_with_q_input argument
