### 2025-03-06 6019b1b
added functionality for flexibly incorporating new embedding networks ; added support for LayerNorm; added support for extra conditioning parameters
