+++ reflectorch/models/networks/mlp_networks.py
@@ -10,75 +10,77 @@ from reflectorch.models.encoders.conv_encoder import ConvEncoder
 from reflectorch.models.encoders.fno import FnoEncoder
 from reflectorch.models.activations import activation_by_name
 
-class NetworkWithPriorsConvEmb(nn.Module):
-    """MLP network with 1D CNN embedding network
+class NetworkWithPriors(nn.Module):
+    """MLP network with an embedding network
 
     .. image:: ../documentation/FigureReflectometryNetwork.png
         :width: 800px
         :align: center
 
     Args:
-        in_channels (int, optional): the number of input channels of the 1D CNN. Defaults to 1.
-        hidden_channels (tuple, optional): list with the number of channels for each layer of the 1D CNN. Defaults to (32, 64, 128, 256, 512).
-        dim_embedding (int, optional): the dimension of the embedding produced by the 1D CNN. Defaults to 128.
-        dim_avpool (int, optional): the type of activation function in the 1D CNN. Defaults to 1.
-        embedding_net_activation (str, optional): the type of activation function in the 1D CNN. Defaults to 'gelu'.
-        use_batch_norm (bool, optional): whether to use batch normalization (in both the 1D CNN and the MLP). Defaults to False.
+        embedding_net_type (str): the type of embedding network, either 'conv' or 'fno'.
+        embedding_net_kwargs (dict): dictionary containing the keyword arguments for the embedding network.
         dim_out (int, optional): the dimension of the output produced by the MLP. Defaults to 8.
+        dim_conditioning_params (int, optional): the dimension of other parameters the network is conditioned on (e.g. for the smearing coefficient dq/q)
         layer_width (int, optional): the width of a linear layer in the MLP. Defaults to 512.
         num_blocks (int, optional): the number of residual blocks in the MLP. Defaults to 4.
         repeats_per_block (int, optional): the number of normalization/activation/linear repeats in a block. Defaults to 2.
         mlp_activation (str, optional):  the type of activation function in the MLP. Defaults to 'gelu'.
+        use_batch_norm (bool, optional): whether to use batch normalization in the MLP. Defaults to True.
+        use_layer_norm (bool, optional): whether to use layer normalization in the MLP (if use_batch_norm is False). Defaults to False.
         dropout_rate (float, optional): dropout rate for each block. Defaults to 0.0.
+        tanh_output (bool, optional): whether to apply a tanh function to the output. Defaults to False.
         use_selu_init (bool, optional): whether to use the special weights initialization for the 'selu' activation function. Defaults to False.
         pretrained_embedding_net (str, optional): the path to the weights of a pretrained embedding network. Defaults to None.
         residual (bool, optional): whether the blocks have a residual skip connection. Defaults to True.
         adaptive_activation (bool, optional): must be set to ``True`` if the activation function is adaptive. Defaults to False.
         conditioning (str, optional): the manner in which the prior bounds are provided as input to the network. Defaults to 'concat'.
-    """
+      """
     def __init__(self,
-                 in_channels: int = 1,
-                 hidden_channels: tuple = (32, 64, 128, 256, 512),
-                 dim_embedding: int = 128,
-                 dim_avpool: int = 1,
-                 embedding_net_activation: str = 'gelu',
-                 use_batch_norm: bool = False,
+                 embedding_net_type: str,  # 'conv', 'fno'
+                 embedding_net_kwargs: dict,
+                 pretrained_embedding_net: str = None,
                  dim_out: int = 8,
+                 dim_conditioning_params: int = 0,
[…]
