### 2025-03-06 06d0bb2
added NetworkWithPriors
