+++ reflectorch/models/encoders/__init__.py
@@ -2,7 +2,6 @@ from reflectorch.models.encoders.conv_encoder import (
     ConvEncoder,
     ConvDecoder,
     ConvAutoencoder,
-    ConvVAE,
 )
 from reflectorch.models.encoders.fno import FnoEncoder, SpectralConv1d
 from reflectorch.models.encoders.conv_res_net import ConvResidualNet1D
@@ -12,7 +11,6 @@ __all__ = [
     "ConvEncoder",
     "ConvDecoder",
     "ConvAutoencoder",
-    "ConvVAE",
     "ConvResidualNet1D",
     "FnoEncoder",
     "SpectralConv1d",
