### 2025-03-06 663143a
included additional arguments; added support for SE blocks in ConvEncoder
