### 2025-03-06 3dd8510
added support for LayerNorm; extended options for conditioning
