+++ reflectorch/ml/schedulers.py
@@ -1,3 +1,4 @@
+import math
 from torch.optim import lr_scheduler
 
 import numpy as np
