### 2025-03-07 ca08885
imporoved functionality for resolution smearing
