### 2025-03-07 96e6139
extended functionality for smearing
