### 2025-03-07 a25337a
added pointwise smearing simulator
