### 2025-03-07 b91fe28
added support for pointwise smearing
