### 2025-03-07 a4758a5
extended get_param_labels for other parameterizations
