### 2025-03-07 96b1f01
backcompatibility with older version of config files
