### 2025-03-07 b8854d1
extended polishing functionality for other parameterizations; added support for kwargs provided to the reflectivity fucntion during polishing
