### 2025-03-07 21dc12b
improved inference class
