### 2025-04-08 c90ea0e
added max_total_thickness argument
