### 2025-04-08 6226285
added r_scale and background as arguments
