### 2025-04-08 74206be
ensure correct shape and dtype for inference
