### 2025-04-08 58e579c
updated noise classes
