### 2025-04-08 fe810fb
updated VariableQ generator
