### 2025-04-08 a1c3544
enable polishing when some parameters are fixed
