### 2025-04-08 3fc667d
compute sld profiles for different ambients
