+++ reflectorch/data_generation/utils.py
@@ -57,7 +57,7 @@ def get_reversed_params(thicknesses: Tensor, roughnesses: Tensor, slds: Tensor):
     return reversed_params
 
 
-def get_density_profiles(
+def get_density_profiles_sld(
         thicknesses: Tensor,
         roughnesses: Tensor,
         slds: Tensor,
@@ -120,6 +120,81 @@ def get_erf(z, z0, sigma, amp):
 def get_gauss(z, z0, sigma, amp):
     return amp / (sigma * sqrt(2 * pi)) * torch.exp(- (z - z0) ** 2 / 2 / sigma ** 2)
 
+def get_density_profiles(
+    thicknesses: torch.Tensor,
+    roughnesses: torch.Tensor,
+    slds: torch.Tensor,
+    ambient_sld: torch.Tensor = None,
+    z_axis: torch.Tensor = None,
+    num: int = 1000,
+    padding_left: float = 0.2,
+    padding_right: float = 1.1,
+):
+    """
+    Args:
+        thicknesses (Tensor): finite layer thicknesses.
+        roughnesses (Tensor): interface roughnesses for all transitions (ambient→layer1 ... layerN→substrate).
+        slds (Tensor): SLDs for the finite layers + substrate.
+        ambient_sld (Tensor, optional): SLD for the top ambient. Defaults to 0.0 if None.
+        z_axis (Tensor, optional): a custom depth axis. If None, a linear axis is generated.
+        num (int): number of points in the generated z-axis (if z_axis is None).
+        padding_left (float): factor to extend the negative (above the surface) portion of z-axis.
+        padding_right (float): factor to extend the positive (into the sample) portion of z-axis.
+
+    Returns:
+        (z_axis, profile, d_profile)
+          z_axis:    1D Tensor of shape (num, ) with the depth coordinates.
+          profile:   2D Tensor of shape (batch_size, num) giving the SLD at each depth.
+          d_profile: 2D Tensor of shape (batch_size, num) giving d(SLD)/dz at each depth.
+    """
+
+    bs, n = thicknesses.shape
+    assert roughnesses.shape == (bs, n + 1),  (
+        f"Roughnesses must be (batch_size, num_layers+1). Found {roughnesses.shape} instead."
+    )
+    assert slds.shape == (bs, n + 1), (
+        f"SLDs must be (batch_size, num_layers+1). Found {slds.shape} instead."
+    )
+    assert torch.all(thicknesses >= 0), "Negative thickness encountered."
+    assert torch.all(roughnesses >= 0), "Negative roughness encountered."
+
+    if ambient_sld is None:
+        ambient_sld = torch.zeros((bs, 1), device=thicknesses.device)
+    else:
+        if ambient_sld.ndim == 1:
+            ambient_sld = ambient_sld.unsqueeze(-1)
+
+    slds_all = torch.cat([ambient_sld, slds], dim=-1)  # new dimension: n+2
+    d_rhos = torch.diff(slds_all, dim=-1)  # (bs, n+1)
+
+    interfaces = torch.cat([
+        torch.zeros((bs, 1), device=thicknesses.device),  # z=0 for ambient→layer1
+        thicknesses
+    ], dim=-1).cumsum(dim=-1)  # now shape => (bs, n+1)
+
+    total_thickness = interfaces[..., -1].max()
+    if z_axis is None:
+        z_axis = torch.linspace(
+            -padding_left * total_thickness,
+            padding_right * total_thickness,
+            num,
+            device=thicknesses.device
+        )  # shape => (num,)
[…]
