### 2025-04-08 2cd562e
added weights_only argument to torch.load for compatibility with the latest Pytorch version
