### 2025-04-08 5663fd4
expanded functionality to print information about the models when loading
