### 2025-04-08 6d0cc53
updated EasyInferenceModel
