+++ reflectorch/inference/inference_model.py
@@ -150,7 +150,8 @@ class EasyInferenceModel(object):
                 reflectivity_curve: Union[np.ndarray, Tensor], 
                 q_values: Union[np.ndarray, Tensor] = None, 
                 prior_bounds: Union[np.ndarray, List[Tuple]] = None, 
-                q_resolution: float = None,
+                q_resolution: Union[float, np.ndarray] = None,
+                ambient_sld: float = None,
                 clip_prediction: bool = False, 
                 polish_prediction: bool = False, 
                 polishing_kwargs_reflectivity: dict = None,
@@ -167,10 +168,11 @@ class EasyInferenceModel(object):
             reflectivity_curve (Union[np.ndarray, Tensor]): The reflectivity curve (which has been already preprocessed, normalized and interpolated).
             q_values (Union[np.ndarray, Tensor], optional): The momentum transfer (q) values for the reflectivity curve (in units of inverse angstroms).
             prior_bounds (Union[np.ndarray, List[Tuple]], optional): the prior bounds for the thin film parameters.
-            q_resolution (float, optional): only for models where the q resolution is an additional input to the network.
+            q_resolution (Union[float, np.ndarray], optional): the instrumental resolution. Either as a float with meaning dq/q for linear smearing or as a numpy array with meaning dq for pointwise smearing. 
+            ambient_sld (float, optional): the SLD of the ambient medium (fronting), if different from air.
             clip_prediction (bool, optional): If ``True``, the values of the predicted parameters are clipped to not be outside the interval set by the prior bounds. Defaults to False.
             polish_prediction (bool, optional): If ``True``, the neural network predictions are further polished using a simple least mean squares (LMS) fit. Only for the standard box-model parameterization. Defaults to False.
-            polishing_kwargs_reflectivity (): extra arguments for the reflectivity function used during polishing. {'dq': ...} must be provided to account for smearing during polishing.
+            polishing_kwargs_reflectivity (dict): extra arguments for the reflectivity function used during polishing.
             fit_growth (bool, optional): If ``True``, an additional parameters is introduced during the LMS polishing to account for the change in the thickness of the upper layer during the in-situ measurement of the reflectivity curve (a linear growth is assumed). Defaults to False.
             max_d_change (float): The maximum possible change in the thickness of the upper layer during the in-situ measurement, relevant when polish_prediction and fit_growth are True. Defaults to 5. 
             use_q_shift: If ``True``, the prediction is performed for a batch of slightly shifted versions of the input curve and the best result is returned, which is meant to mitigate the influence of imperfect sample alignment, as introduced in Greco et al. (only for models with fixed q-discretization). Defaults to False.
@@ -184,6 +186,17 @@ class EasyInferenceModel(object):
 
         scaled_curve = self._scale_curve(reflectivity_curve)
         prior_bounds = np.array(prior_bounds)
+
+        if ambient_sld:
+            n_layers = self.trainer.loader.prior_sampler.max_num_layers
+            sld_indices = slice(2*n_layers+1, 3*n_layers+2)
+            prior_bounds[sld_indices, ...] -= ambient_sld
+            training_min_bounds = self.trainer.loader.prior_sampler.min_bounds.squeeze().cpu().numpy()
+            training_max_bounds = self.trainer.loader.prior_sampler.max_bounds.squeeze().cpu().numpy()
+            lower_bound_check = (prior_bounds[sld_indices, 0] >= training_min_bounds[sld_indices]).all()
+            upper_bound_check = (prior_bounds[sld_indices, 1] <= training_max_bounds[sld_indices]).all()
[…]
