+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "reflectorch"
-version = "2.0.0"
+version = "1.3.0"
 authors = [{name = "Vladimir Starostin", email = "vladimir.starostin@uni-tuebingen.de"},
            {name = "Valentin Munteanu", email = "valentin.munteanu@uni-tuebingen.de"}
 ]
@@ -29,7 +29,7 @@ classifiers = [
 ]
 requires-python = ">=3.7"
 dependencies = [
-    "numpy>=1.18.1,<2.0",
+    "numpy", #"numpy>=1.18.1,<2.0",
     "torch>=1.8.1",
     "scipy",
     "tqdm",
@@ -39,6 +39,7 @@ dependencies = [
     "ipywidgets",
     "huggingface_hub",
     "safetensors",
+    "tensorboard"
 ]
 
 [project.optional-dependencies]
