+++ pyproject.toml
@@ -29,7 +29,7 @@ classifiers = [
 ]
 requires-python = ">=3.7"
 dependencies = [
-    "numpy", #"numpy>=1.18.1,<2.0",
+    "numpy",
     "torch>=1.8.1",
     "scipy",
     "tqdm",
