### 2025-04-24 c10f65f
added support for solvent volume fraction
