+++ reflectorch/data_generation/reflectivity/__init__.py
@@ -26,6 +26,8 @@ def reflectivity(
         q_shift: Tensor = 0.0,
         r_scale: Tensor = 1.0,
         background: Tensor = 0.0,
+        solvent_vf = None,
+        solvent_mode = 'fronting',
         abeles_func = None,
         **abeles_kwargs
 ):
@@ -55,6 +57,21 @@ def reflectivity(
     abeles_func = abeles_func or abeles
     q = torch.atleast_2d(q) + q_shift
     q = torch.clamp(q, min=0.0)
+    
+    if solvent_vf is not None:
+        num_layers = thickness.shape[-1]
+        if solvent_mode == 'fronting':
+            assert sld.shape[-1] == num_layers + 2
+            assert solvent_vf.shape[-1] == num_layers
+            solvent_sld = sld[..., [0]]
+            idx = slice(1, num_layers)
+            sld[..., idx] = solvent_vf * solvent_sld + (1.0 - solvent_vf) * sld[..., idx]
+        elif solvent_mode == 'backing':
+            solvent_sld = sld[..., [-1]]
+            idx = slice(1, num_layers) if sld.shape[-1] == num_layers + 2 else slice(0, num_layers)
+            sld[..., idx] = solvent_vf * solvent_sld + (1.0 - solvent_vf) * sld[..., idx]
+        else:
+            raise NotImplementedError
 
     if dq is None:
         reflectivity_curves = abeles_func(q, thickness, roughness, sld, **abeles_kwargs)
