### 2025-04-24 7d2024e
updated EasyInferenceModel
