+++ reflectorch/inference/inference_model.py
@@ -12,8 +12,10 @@ from IPython.display import display
 from huggingface_hub import hf_hub_download
 
 from reflectorch.data_generation.priors import Params, BasicParams, ExpUniformSubPriorSampler, UniformSubPriorParams
+from reflectorch.data_generation.priors.parametric_models import NuisanceParamsWrapper
 from reflectorch.data_generation.q_generator import ConstantQ, VariableQ
 from reflectorch.data_generation.utils import get_density_profiles, get_param_labels
+from reflectorch.inference.plotting import plot_prediction_results
 from reflectorch.inference.preprocess_exp.interpolation import interp_reflectivity
 from reflectorch.paths import CONFIG_DIR, ROOT_DIR, SAVED_MODELS_DIR
 from reflectorch.runs.utils import (
@@ -104,7 +106,9 @@ class EasyInferenceModel(object):
         self.trainer = get_trainer_by_name(config_name=config_name, config_dir=self.config_dir, model_path=model_path, load_weights=True, inference_device = self.device)
         self.trainer.model.eval()
         
-        print(f'The model corresponds to a `{self.trainer.loader.prior_sampler.param_model.NAME}` parameterization with {self.trainer.loader.prior_sampler.max_num_layers} layers ({self.trainer.loader.prior_sampler.param_dim} predicted parameters)')
+        param_model = self.trainer.loader.prior_sampler.param_model
+        param_model_name = param_model.base_model.NAME if isinstance(param_model, NuisanceParamsWrapper) else param_model.NAME
+        print(f'The model corresponds to a `{param_model_name}` parameterization with {self.trainer.loader.prior_sampler.max_num_layers} layers ({self.trainer.loader.prior_sampler.param_dim} predicted parameters)')
         print("Parameter types and total ranges:")
         for param, range_ in self.trainer.loader.prior_sampler.param_ranges.items():
             print(f"- {param}: {range_}")
@@ -286,9 +290,8 @@ class EasyInferenceModel(object):
 
         return prediction_dict
 
-    async def predict_using_widget(self, reflectivity_curve: Union[np.ndarray, torch.Tensor], **kwargs):
-        """Use an interactive Python widget for specifying the prior bounds before the prediction (works only in a Jupyter notebook).
-        The other arguments are the same as for the ``predict`` method.
+    def predict_using_widget(self, reflectivity_curve, **kwargs):
+        """
         """
 
         NUM_INTERVALS = self.trainer.loader.prior_sampler.param_dim
@@ -297,76 +300,74 @@ class EasyInferenceModel(object):
         max_bounds = self.trainer.loader.prior_sampler.max_bounds.cpu().numpy().flatten()
         max_deltas = self.trainer.loader.prior_sampler.max_delta.cpu().numpy().flatten()
 
-        print(f'Parameter ranges: {self.trainer.loader.prior_sampler.param_ranges}')
-        print(f'Allowed widths of the prior bound intervals (max-min): {self.trainer.loader.prior_sampler.bound_width_ranges}')
-        print(f'Please fill in the values of the minimum and maximum prior bound for each parameter and press the button!')
+        print(f'Adjust the sliders for each parameter and press "Predict". Repeat as desired. Press "Close Widget" to finish.')
 
-        def create_interval_widgets(n):
-            intervals = []
-            for i in range(n):
-                interval_label = widgets.Label(value=f'{param_labels[i]}')
-                initial_max = min(max_bounds[i], min_bounds[i] + max_deltas[i])
-                slider = widgets.FloatRangeSlider(
-                    value=[min_bounds[i], initial_max],
-                    min=min_bounds[i],
-                    max=max_bounds[i],
-                    step=0.01,
-                    layout=widgets.Layout(width='400px'),
[…]
