+++ documentation/api.rst
@@ -26,6 +26,7 @@ API Reference
               OneCycleLR,
               CyclicLR,
               LogCyclicLR,
+              CosineAnnealingWithWarmup,
               QGenerator,
               ConstantQ, 
               VariableQ,
@@ -37,19 +38,16 @@ API Reference
               QNoiseGenerator,
               BasicQNoiseGenerator,
               IntensityNoiseGenerator,
-              MultiplicativeLogNormalNoiseGenerator,
-              PoissonNoiseGenerator,
-              BasicExpIntensityNoise,
+              GaussianNoiseGenerator,
+              GaussianExpIntensityNoise,
               Smearing,
-              angle_to_q, 
-              q_to_angle, 
-              energy_to_wavelength, 
-              wavelength_to_energy,
               get_density_profiles,
               ReflectivityDataLoader,
               Logger,
               Loggers,
               PrintLogger,
+              TensorBoardLogger,
+              NetworkWithPriors,
               NetworkWithPriorsConvEmb, 
               NetworkWithPriorsFnoEmb,
               ConvEncoder,
@@ -58,7 +56,6 @@ API Reference
               apply_footprint_correction,
               interp_reflectivity,
               standard_preprocessing,
-              ReflGradientFit,
               reflectivity,
               LogLikelihood,
               PoissonLogLikelihood,
