+++ documentation/api.rst
@@ -38,7 +38,6 @@ API Reference
               QNoiseGenerator,
               BasicQNoiseGenerator,
               IntensityNoiseGenerator,
-              GaussianNoiseGenerator,
               GaussianExpIntensityNoise,
               Smearing,
               get_density_profiles,
@@ -58,7 +57,6 @@ API Reference
               standard_preprocessing,
               reflectivity,
               LogLikelihood,
-              PoissonLogLikelihood,
               Params,
               BasicParams,
               PriorSampler,
