### 2025-05-05 204115b
fixed safetensors conversion for latest pytorch version
