+++ configs/b_mc_point_neutron_conv_standard_L1_InputDq.yaml
@@ -1,0 +1,118 @@
+general:
+  name: b_mc_point_neutron_conv_standard_L1_InputDq
+  root_dir: null
+  
+dset:
+  cls: ReflectivityDataLoader
+  prior_sampler:
+    cls: SubpriorParametricSampler
+    kwargs:
+      param_ranges:
+        thicknesses: [1., 1000.]
+        roughnesses: [0., 60.]
+        slds: [-8., 16.]
+      bound_width_ranges:
+        thicknesses: [1.0e-2, 1000.]
+        roughnesses: [1.0e-2, 60.]
+        slds: [1.0e-2, 5.]
+      model_name: standard_model
+      max_num_layers: 1
+      constrained_roughness: true
+      max_thickness_share: 0.5
+      logdist: false
+      scale_params_by_ranges: false
+      scaled_range: [-1., 1.]
+      device: 'cuda'
+      
+  q_generator:
+    cls: ConstantQ
+    kwargs:
+      q: [0.005, 0.2, 128]
+      device: 'cuda'
+
+  intensity_noise:
+    cls: GaussianExpIntensityNoise
+    kwargs:
+      relative_errors: [0.01, 0.3]
+      consistent_rel_err: false
+      apply_shift: true
+      shift_range: [-0.3, 0.3]
+      apply_background: true
+      background_range: [1.0e-10, 1.0e-4]
+      add_to_context: true
+
+  smearing:
+    cls: Smearing
+    kwargs:
+      sigma_range: [0.01, 0.10]
+      constant_dq: False
+      gauss_num: 17
+      share_smeared: 0.8
+      
+  curves_scaler:
+    cls: LogAffineCurvesScaler
+    kwargs:
+      weight: 0.2
+      bias: 1.0
+      eps: 1.0e-10
+
+model:
+  network:
+    cls: NetworkWithPriors
+    pretrained_name: null
+    device: 'cuda'
+    kwargs:
+      embedding_net_type: 'conv'
+      embedding_net_kwargs:
+        in_channels: 1
+        hidden_channels: [32, 64, 128, 256, 512]
+        kernel_size: 3
+        dim_embedding: 128
+        dim_avpool: 1
+        use_batch_norm: true
+        use_se: false
+        activation: 'gelu'
+      pretrained_embedding_net: null
+      dim_out: 5
+      dim_conditioning_params: 1
+      layer_width: 512
+      num_blocks: 8
+      repeats_per_block: 2
+      residual: true
+      use_batch_norm: true
+      use_layer_norm: false
+      mlp_activation: 'gelu'
+      dropout_rate: 0.0 
+      tanh_output: false
+      conditioning: 'film'
+      concat_condition_first_layer: true ### back. comp.
+       
+training:
+  trainer_cls: PointEstimatorTrainer
+  num_iterations: 100000
+  batch_size: 4096
+  lr: 1.0e-3
+  grad_accumulation_steps: 1
+  clip_grad_norm_max: null
+  update_tqdm_freq: 1
+  optimizer: AdamW
+  trainer_kwargs:
+    train_with_q_input: false
+    train_with_sigmas: false
+    condition_on_q_resolutions: true
+    use_curve_reconstruction_loss: false
+    rescale_loss_interval_width: true
+    use_l1_loss: true
+    optim_kwargs:
+      betas: [0.9, 0.999]
+      weight_decay: 0.0005
+  callbacks:
+    save_best_model:
+      enable: true
+      freq: 500
+    lr_scheduler:
+      cls: CosineAnnealingWithWarmup
+      kwargs:
+        min_lr: 1.0e-6
+        warmup_iters: 500
+        total_iters: 100000
