### 2025-05-06 2389ae0
added check for compatibiltity between number of provided and expected prior bounds
