### 2025-05-06 #7089 @turnerjennings
> 👋 [@turnerjennings](https://github.com/turnerjennings) - Would you be willing to review this submission to JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

@crvernon Happy to review this submission!  This is my first review with JOSS so bear with me while I figure out the format.
