### 2025-05-06 #7089 @rachitanaik-lyft
Hi @crvernon! I don't have domain knowledge about X-ray and neutron reflectometry - is that a requirement?
