### 2025-05-06 #8169 @crvernon
👋 @valentinsingularity , @turnerjennings , @thomashopkins32 , and @rachitanaik-lyft  - This is the review thread for the paper. All of our communications will happen here from now on.

I'll turn this back over to your topic editor @likeajumprope now.
