### 2025-05-08 #8169 @likeajumprope
> 👋 [@valentinsingularity](https://github.com/valentinsingularity) , [@turnerjennings](https://github.com/turnerjennings) , [@thomashopkins32](https://github.com/thomashopkins32) , and [@rachitanaik-lyft](https://github.com/rachitanaik-lyft) - This is the review thread for the paper. All of our communications will happen here from now on.
>
> Please read the "Reviewer instructions & questions" in the first comment above.
>
> Both reviewers have checklists at the top of this thread (in that first comment) with the JOSS requirements. As you go over the submission, please check any items that you feel have been satisfied. There are also links to the JOSS reviewer guidelines.
>
> The JOSS review is different from most other journals. Our goal is to work with the authors to help them meet our criteria instead of merely passing judgment on the submission. As such, the reviewers are encouraged to submit issues and pull requests on the software repository. When doing so, please mention [#8169](https://github.com/openjournals/joss-reviews/issues/8169) so that a link is created to this thread (and I can keep an eye on what is happening). Please also feel free to comment and ask questions on this thread. In my experience, it is better to post comments/questions/suggestions as you come across them instead of waiting until you've reviewed the entire package.
>
> We aim for the review process to be completed within about 4-6 weeks but please make a start well ahead of this as JOSS reviews are by their nature iterative and any early feedback you may be able to provide to the author will be very helpful in meeting this schedule.
>
> I'll turn this back over to your topic editor [@likeajumprope](https://github.com/likeajumprope) now.

thank you for moving this forward @crvernon!
