### 2025-05-13 #8169 @thomashopkins32
> Repository: Is the source code for this software available at the https://github.com/schreiber-lab/reflectorch?

The source code is available, however, it may be helpful to have tags and releases in the GitHub repository for the versions available at https://pypi.org/project/reflectorch/. If we are meant to review specifically v1.2.1 then it's difficult to find the source code for that specific release. It also helps if I am a user of this software to be able to reference a specific tag/commit hash in a GitHub Issue or Pull Request. I may not need to raise an issue at all if I can easily compare the source code for version v1.2.1 with what is in the `master` branch and notice that the issue is resolved already.
