### 2025-05-14 #8169 @valentinsingularity
@thomashopkins32  Since our initial submission in August 2024 we have made substantial improvements to the software which are integrated in the latest version (v1.3.0). This version aligns with the current documentation, while v.1.2.1 is partly outdated. As such, we kindly request that the latest version be considered for review.
