### 2025-05-20 schreiber-lab/reflectorch#22 @thomashopkins32
Docs Issue: Using a trained reflectorch model missing some code
I think section 3 of the docs may be missing some code. I've been trying to work through the examples as a reviewer for https://github.com/openjournals/joss-reviews/issues/8169 and I found that the plots after
```
predicted_curves = restored_predictions.reflectivity(q)
```
were not generated and the code that produces them doesn't seem to be present. Trying to run the final hidden code block also does not work.

Here is the section I am referring to: https://schreiber-lab.github.io/reflectorch/using_reflectorch.html#applying-the-model-to-synthetic-data

The code appears here in the docs `.ipynb` so maybe it's a build issue? https://github.com/schreiber-lab/reflectorch/blob/master/docs/_sources/using_reflectorch.ipynb

I find it useful to strip the output of the notebooks using [pre-commit](https://pre-commit.com/) and https://github.com/kynan/nbstripout and have sphinx run and render the notebooks.
