+++ pyproject.toml
@@ -17,14 +17,14 @@ maintainers = [
 description = "A Pytorch-based package for the analysis of reflectometry data"
 keywords = ["reflectometry", "machine learning"]
 readme = "README.md"
-license = "MIT"
+license = {file = "LICENSE"}
 classifiers = [
     "Programming Language :: Python :: 3",
     "Operating System :: OS Independent",
     "Environment :: GPU :: NVIDIA CUDA",
     "Development Status :: 4 - Beta",
     "Topic :: Scientific/Engineering :: Physics",
-    "Intended Audience :: Science/Research"
+    "Intended Audience :: Science/Research",
 ]
 requires-python = ">=3.7"
 dependencies = [
