### 2025-05-22 #8169 @valentinsingularity
We have slightly shortened the 'General workflow' section of the paper.

@thomashopkins32  The references file can be found on the master branch at the following location: https://github.com/schreiber-lab/reflectorch/blob/master/paper/paper.bib

I have added a CONTRIBUTING.md file to the repository, which is also referenced in the README.md file.
