### 2025-08-19 023283f
added the IntegralConv embedding network
