### 2025-08-19 eaa2134
updated ParametricModel
