### 2025-08-19 b2d17c5
updated get_density_profiles
