### 2025-08-19 e76879b
added MaskedVariableQ generator
