### 2025-08-19 ce608c2
updated polishing functionality
