### 2025-08-19 ce8641b
plotting utilities
