+++ reflectorch/inference/plotting.py
@@ -1,98 +1,517 @@
-from matplotlib import pyplot as plt
 import numpy as np
+import matplotlib.pyplot as plt
+import matplotlib.ticker as mticker
+from matplotlib.lines import Line2D
+
+
+def print_prediction_results(prediction_dict, param_names=None, width=10, precision=3, header=True):
+
+    if param_names is None:
+        param_names = prediction_dict.get("param_names", [])
+
+    pred = np.asarray(prediction_dict.get("predicted_params_array", []), dtype=float)
+    pol = prediction_dict.get("polished_params_array", None)
+    pol = np.asarray(pol, dtype=float) if pol is not None else None
+
+    name_w = max(14, max((len(str(n)) for n in param_names), default=14))
+
+    num_fmt = f"{{:>{width}.{precision}f}}"
+
+    if header:
+        hdr = f"{'Parameter'.ljust(name_w)}  {'Predicted'.rjust(width)}"
+        if pol is not None:
+            hdr += f"  {'Polished'.rjust(width)}"
+        print(hdr)
+        print("-" * len(hdr))
+
+    for i, name in enumerate(param_names):
+        pred_val = pred[i] if i < pred.size else float("nan")
+        row = f"{str(name).ljust(name_w)}  {num_fmt.format(pred_val)}"
+        if pol is not None:
+            pol_val = pol[i] if i < pol.size else float("nan")
+            row += f"  {num_fmt.format(pol_val)}"
+        print(row)
 
 
 def plot_prediction_results(
     prediction_dict: dict,
-    q_exp: np.ndarray = None,
-    curve_exp: np.ndarray = None,
+    q_exp: np.ndarray,
+    curve_exp: np.ndarray,
     sigmas_exp: np.ndarray = None,
-    q_model: np.ndarray = None,
+    logx=False,
 ):
-    """
-    Plot the experimental curve (with optional error bars), the predicted
-    and polished curves, and also the predicted/polished SLD profiles.
+    q_pred = prediction_dict['q_plot_pred']
+    r_pred = prediction_dict['predicted_curve']
+    r_pol = prediction_dict.get('polished_curve', None)
 
-    Args:
-        prediction_dict (dict): Dictionary containing 'predicted_curve',
-                                'predicted_sld_profile', 'predicted_sld_xaxis',
-                                and optionally 'polished_curve', 'sld_profile_polished'.
-        q_exp (ndarray, optional): Experimental q-values.
-        curve_exp (ndarray, optional): Experimental reflectivity curve.
-        sigmas_exp (ndarray, optional): Error bars of the experimental reflectivity.
-        q_model (ndarray, optional): The q-values on which prediction_dict's reflectivity
-                                     was computed (e.g. from EasyInferenceModel.interpolate_data_to_model_q).
+    q_pol = None
+    if r_pol is not None:
+        if len(r_pol) == len(q_pred):
+            q_pol = q_pred
+        elif len(r_pol) == len(q_exp):
+            q_pol = q_exp
 
-    Example usage:
-        prediction_dict = model.predict(...)
-        plot_prediction_results(
-            prediction_dict,
-            q_exp=q_exp,
-            curve_exp=curve_exp,
-            sigmas_exp=sigmas_exp,
-            q_model=q_model
-        )
-    """
+    z_sld = prediction_dict.get('predicted_sld_xaxis', None)
[…]
