### 2025-08-19 9994ff6
updated EasyInferenceModel
