+++ reflectorch/inference/inference_model.py
@@ -13,9 +13,8 @@ from huggingface_hub import hf_hub_download
 
 from reflectorch.data_generation.priors import Params, BasicParams, ExpUniformSubPriorSampler, UniformSubPriorParams
 from reflectorch.data_generation.priors.parametric_models import NuisanceParamsWrapper
-from reflectorch.data_generation.q_generator import ConstantQ, VariableQ
+from reflectorch.data_generation.q_generator import ConstantQ, VariableQ, MaskedVariableQ
 from reflectorch.data_generation.utils import get_density_profiles, get_param_labels
-from reflectorch.inference.plotting import plot_prediction_results
 from reflectorch.inference.preprocess_exp.interpolation import interp_reflectivity
 from reflectorch.paths import CONFIG_DIR, ROOT_DIR, SAVED_MODELS_DIR
 from reflectorch.runs.utils import (
@@ -29,7 +28,8 @@ from reflectorch.inference.preprocess_exp import StandardPreprocessing
 from reflectorch.inference.scipy_fitter import standard_refl_fit, refl_fit, get_fit_with_growth
 from reflectorch.inference.sampler_solution import simple_sampler_solution, get_best_mse_param
 from reflectorch.inference.record_time import print_time
-from reflectorch.utils import to_t
+from reflectorch.inference.plotting import plot_reflectivity, plot_prediction_results, print_prediction_results
+from reflectorch.utils import get_filtering_mask, to_t
 
 class EasyInferenceModel(object):
     """Facilitates the inference process using pretrained models
@@ -150,21 +150,158 @@ class EasyInferenceModel(object):
             inputs_str = ", ".join(additional_inputs)
             print(f"The following quantities are additional inputs to the network: {inputs_str}.")
 
+    def preprocess_and_predict(self, 
+                            reflectivity_curve: Union[np.ndarray, Tensor], 
+                            q_values: Union[np.ndarray, Tensor] = None, 
+                            prior_bounds: Union[np.ndarray, List[Tuple]] = None, 
+                            sigmas: Union[np.ndarray, Tensor] = None,
+                            q_resolution: float = None,
+                            ambient_sld: float = None,
+                            clip_prediction: bool = True, 
+                            polish_prediction: bool = False,
+                            polishing_method: str = 'trf',
+                            polishing_kwargs_reflectivity: dict = None,
+                            use_sigmas_for_polishing: bool = False,
+                            polishing_max_nfev: int = None,
+                            fit_growth: bool = False, 
+                            max_d_change: float = 5.,
+                            calc_pred_curve: bool = True,
+                            calc_pred_sld_profile: bool = False,
+                            calc_polished_sld_profile: bool = False,
+                            sld_profile_padding_left: float = 0.2,
+                            sld_profile_padding_right: float = 1.1,
+                            kwargs_param_labels: dict = {},
+                            
+                            truncate_index_left: int = None,
+                            truncate_index_right: int = None,
+                            enable_error_bars_filtering: bool = True,
+                            filter_threshold=0.3,
+                            filter_remove_singles=True,
+                            filter_remove_consecutives=True,
+                            filter_consecutive=3,
+                            filter_q_start_trunc=0.1,
+                            ):
+        
+        ## Preprocess the data for inference (remove negative intensities, truncation, filer out points with high error bars)
+        (q_values, reflectivity_curve, sigmas, q_resolution, 
+         q_values_original, reflectivity_curve_original, sigmas_original, q_resolution_original) = self._preprocess_input_data(
[…]
