### 2025-08-28 553af3b
selected changes to incorporate logspace functionality in VariableQ generator class
