+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "reflectorch"
-version = "1.3.0"
+version = "1.4.0"
 authors = [{name = "Vladimir Starostin", email = "vladimir.starostin@uni-tuebingen.de"},
            {name = "Valentin Munteanu", email = "valentin.munteanu@uni-tuebingen.de"}
 ]
