### 2025-09-08 ac5cf84
separating widget implementation
