+++ reflectorch/extensions/jupyter/ui.py
@@ -1,0 +1,495 @@
+
+
+"""
+Jupyter Widget Interface for Reflectorch
+
+This module provides interactive Jupyter widgets for reflectometry analysis using Reflectorch models.
+The widgets offer a user-friendly interface for data preprocessing, prediction, and visualization.
+
+Key Features:
+    - Interactive parameter bounds adjustment with sliders
+    - Real-time prediction and visualization
+    - Comprehensive data preprocessing controls
+    - Customizable plotting and color options
+    - Easy integration with existing Reflectorch workflows
+
+Quick Start:
+    ```python
+    # Basic widget usage
+    from reflectorch.inference import EasyInferenceModel
+    from reflectorch.extensions.jupyter import create_basic_widget
+    
+    model = EasyInferenceModel('your_config.yaml')
+    widget = create_basic_widget(model, reflectivity_curve, q_values=q_values)
+    
+    # Advanced widget usage
+    from reflectorch.extensions.jupyter import create_advanced_widget
+    
+    widget = create_advanced_widget(
+        model, 
+        reflectivity_curve, 
+        q_values,
+        sigmas=sigmas,
+        initial_prior_bounds=my_bounds
+    )
+    
+    # Access results
+    results = widget.prediction_result
+    ```
+
+Widget Types:
+    1. **Basic Widget**: Simple interface for adjusting prior bounds and making predictions
+    2. **Advanced Widget**: Full-featured interface with preprocessing and visualization controls
+"""
+
+import numpy as np
+from typing import Optional, Dict, Any, Union, List, Tuple
+import ipywidgets as widgets
+from IPython.display import display
+
+from ...inference.plotting import plot_reflectivity, plot_prediction_results, print_prediction_results
+
+
+class ReflectJupyterWidget:
+    """
+    Interactive Jupyter Widget for Reflectometry Analysis with Reflectorch
+    
+    This class provides two main widget interfaces:
+    1. Basic prediction widget with prior bounds adjustment
+    2. Advanced preprocessing and prediction widget with full control
+    
+    The widgets allow interactive adjustment of parameters and real-time visualization
+    of prediction results.
+    
+    Attributes:
+        model: The EasyInferenceModel instance
+        prediction_result (dict): Latest prediction results from widget interaction
+        
+    Example:
+        ```python
+        from reflectorch.inference import EasyInferenceModel
+        from reflectorch.extensions.jupyter import ReflectJupyterWidget
+        
+        # Load model
+        model = EasyInferenceModel('your_config')
+        
+        # Create widget
+        widget = ReflectJupyterWidget(model)
+        
+        # Use basic prediction widget
+        widget.basic_predict_widget(reflectivity_curve, q_values=q_values)
+        
+        # Or use advanced widget with preprocessing
+        widget.advanced_predict_widget(
+            reflectivity_curve, 
+            q_values=q_values,
+            sigmas=sigmas
+        )
+        ```
+    """
+    
+    def __init__(self, model):
+        """
+        Initialize the widget with a Reflectorch model
+        
+        Args:
+            model: EasyInferenceModel instance for making predictions
+        """
+        self.model = model
+        self.prediction_result = None
+        self._validate_model()
+    
+    def _validate_model(self):
+        """Validate that the model has required attributes"""
+        required_attrs = ['trainer', 'predict', 'preprocess_and_predict']
+        for attr in required_attrs:
+            if not hasattr(self.model, attr):
+                raise ValueError(f"Model must have '{attr}' attribute")
+    
+    def basic_predict_widget(
+        self, 
+        reflectivity_curve: np.ndarray,
+        q_values: Optional[np.ndarray] = None,
+        initial_prior_bounds: Optional[np.ndarray] = None,
+        **predict_kwargs
[…]
