+++ reflectorch/inference/inference_model.py
@@ -7,8 +7,6 @@ import numpy as np
 import torch
 from torch import Tensor
 from typing import List, Tuple, Union
-import ipywidgets as widgets
-from IPython.display import display
 from huggingface_hub import hf_hub_download
 
 from reflectorch.data_generation.priors import Params, BasicParams, ExpUniformSubPriorSampler, UniformSubPriorParams
@@ -765,287 +763,7 @@ class EasyInferenceModel(object):
             )
             return best_param
 
-    def predict_using_widget(self, reflectivity_curve, **kwargs):
-        """
-        """
 
-        NUM_INTERVALS = self.trainer.loader.prior_sampler.param_dim
-        param_labels = self.trainer.loader.prior_sampler.param_model.get_param_labels()
-        min_bounds = self.trainer.loader.prior_sampler.min_bounds.cpu().numpy().flatten()
-        max_bounds = self.trainer.loader.prior_sampler.max_bounds.cpu().numpy().flatten()
-        max_deltas = self.trainer.loader.prior_sampler.max_delta.cpu().numpy().flatten()
-
-        print(f'Adjust the sliders for each parameter and press "Predict". Repeat as desired. Press "Close Widget" to finish.')
-
-        interval_widgets = []
-        for i in range(NUM_INTERVALS):
-            label = widgets.Label(value=f'{param_labels[i]}')
-            initial_max = min(max_bounds[i], min_bounds[i] + max_deltas[i])
-            slider = widgets.FloatRangeSlider(
-                value=[min_bounds[i], initial_max],
-                min=min_bounds[i],
-                max=max_bounds[i],
-                step=0.01,
-                layout=widgets.Layout(width='400px'),
-                style={'description_width': '60px'}
-            )
-
-            def validate_range(change, slider=slider, max_width=max_deltas[i]):
-                min_val, max_val = change['new']
-                if max_val - min_val > max_width:
-                    old_min_val, old_max_val = change['old']
-                    if abs(old_min_val - min_val) > abs(old_max_val - max_val):
-                        max_val = min_val + max_width
-                    else:
-                        min_val = max_val - max_width
-                    slider.value = [min_val, max_val]
-
-            slider.observe(validate_range, names='value')
-            interval_widgets.append((slider, widgets.HBox([label, slider])))
-
-        sliders_box = widgets.VBox([iw[1] for iw in interval_widgets])
-
-        output = widgets.Output()
-        predict_button = widgets.Button(description="Predict")
-        close_button = widgets.Button(description="Close Widget")
-
-        container = widgets.VBox([sliders_box, widgets.HBox([predict_button, close_button]), output])
-        display(container)
-
-        @output.capture(clear_output=True)
-        def on_predict_click(_):
-            if 'prior_bounds' in kwargs:
-                array_values = kwargs.pop('prior_bounds')
-                for i, (s, _) in enumerate(interval_widgets):
-                    s.value = tuple(array_values[i])
-            else:
-                values = [(s.value[0], s.value[1]) for s, _ in interval_widgets]
-                array_values = np.array(values)
-
-            prediction_result = self.predict(reflectivity_curve=reflectivity_curve,
-                                            prior_bounds=array_values,
-                                            **kwargs)
-            param_names = self.trainer.loader.prior_sampler.param_model.get_param_labels()
-            print_prediction_results(prediction_result)
-
-            plot_prediction_results(
-                prediction_result,
-                q_exp=kwargs['q_values'],
-                curve_exp=reflectivity_curve,
-            )
-            self.widget_prediction_result = prediction_result
-
[…]
