### 2025-09-09 c7ab7a1
interactive widget
