+++ reflectorch/extensions/jupyter/ui.py
@@ -8,21 +8,30 @@ The widgets offer a user-friendly interface for data preprocessing, prediction, 
 
 Key Features:
     - Interactive parameter bounds adjustment with sliders
-    - Real-time prediction and visualization
+    - Real-time prediction and visualization with interactive plots
+    - Persistent figures that update in-place (no figure accumulation)
+    - Cursor coordinate display and zoom/pan functionality
     - Comprehensive data preprocessing controls
     - Customizable plotting and color options
     - Easy integration with existing Reflectorch workflows
 
+Interactive Plotting:
+    - Uses ipympl backend for enhanced interactivity in JupyterLab
+    - Figures update in-place rather than creating new windows
+    - Real-time coordinate display as you move cursor over plots
+    - Built-in zoom, pan, and navigation controls
+    - Graceful fallback to standard matplotlib if interactive mode disabled
+
 Quick Start:
     ```python
-    # Basic widget usage
+    # Basic widget usage with interactive plotting
     from reflectorch.inference import EasyInferenceModel
     from reflectorch.extensions.jupyter import create_basic_widget
     
     model = EasyInferenceModel('your_config.yaml')
     widget = create_basic_widget(model, reflectivity_curve, q_values=q_values)
     
-    # Advanced widget usage
+    # Advanced widget usage with full interactive features
     from reflectorch.extensions.jupyter import create_advanced_widget
     
     widget = create_advanced_widget(
@@ -30,8 +39,12 @@ Quick Start:
         reflectivity_curve, 
         q_values,
         sigmas=sigmas,
-        initial_prior_bounds=my_bounds
+        initial_prior_bounds=my_bounds,
+        interactive_plots=True  # Default, enables interactive plotting
     )
+    
+    # Disable interactive plotting if needed
+    widget = create_basic_widget(model, data, interactive_plots=False)
     
     # Access results
     results = widget.prediction_result
@@ -40,6 +53,8 @@ Quick Start:
 Widget Types:
     1. **Basic Widget**: Simple interface for adjusting prior bounds and making predictions
     2. **Advanced Widget**: Full-featured interface with preprocessing and visualization controls
+
+Note: For best interactive experience, ensure ipympl is installed and use '%matplotlib widget' in Jupyter.
 """
 
 import numpy as np
@@ -48,6 +63,7 @@ import ipywidgets as widgets
 from IPython.display import display
 
 from ...inference.plotting import plot_reflectivity, plot_prediction_results, print_prediction_results
+from .interactive_plotting import InteractivePlotManager, plot_reflectivity_interactive, plot_prediction_results_interactive
 
 
 class ReflectJupyterWidget:
@@ -88,15 +104,18 @@ class ReflectJupyterWidget:
         ```
     """
     
-    def __init__(self, model):
+    def __init__(self, model, interactive_plots=True):
         """
         Initialize the widget with a Reflectorch model
         
         Args:
             model: EasyInferenceModel instance for making predictions
+            interactive_plots: Enable interactive plotting features (default: True)
         """
         self.model = model
         self.prediction_result = None
+        self.interactive_plots = interactive_plots
+        self.plot_manager = InteractivePlotManager() if interactive_plots else None
         self._validate_model()
     
     def _validate_model(self):
@@ -149,10 +168,32 @@ class ReflectJupyterWidget:
         
         print(f'Adjust the sliders for each parameter and press "Predict". Repeat as desired. Press "Close Widget" to finish.')
         
-        # Create parameter sliders
-        sliders = self._create_parameter_sliders(
-            param_labels, min_bounds, max_bounds, max_deltas, initial_prior_bounds
-        )
[…]
