+++ reflectorch/extensions/jupyter/__init__.py
@@ -1,6 +1,13 @@
 from .callbacks import JPlotLoss
-
+from reflectorch.extensions.jupyter.ui import create_basic_widget, create_advanced_widget, ReflectJupyterWidget
+from .interactive_plotting import InteractivePlotManager, plot_reflectivity_interactive, plot_prediction_results_interactive
 
 __all__ = [
     'JPlotLoss',
+    'create_basic_widget',
+    'create_advanced_widget',
+    'ReflectJupyterWidget',
+    'InteractivePlotManager',
+    'plot_reflectivity_interactive',
+    'plot_prediction_results_interactive',
 ]
