+++ reflectorch/extensions/jupyter/interactive_plotting.py
@@ -1,0 +1,519 @@
+"""
+Interactive Plotting Support for Reflectorch Jupyter Widgets
+
+This module provides enhanced plotting capabilities with interactive features:
+- Figure persistence and in-place updates
+- Cursor coordinate display
+- Zoom and pan functionality
+- Dynamic plot updates without figure recreation
+
+The module is designed to be a drop-in replacement for static matplotlib plotting
+while adding interactive capabilities for Jupyter environments.
+"""
+
+import numpy as np
+import matplotlib.pyplot as plt
+import matplotlib.ticker as mticker
+from matplotlib.lines import Line2D
+from matplotlib.widgets import Cursor
+from IPython.display import display, HTML
+import ipywidgets as widgets
+from typing import Optional, Tuple, Dict, Any, Union
+import warnings
+from contextlib import contextmanager
+
+try:
+    from IPython import get_ipython
+except ImportError:
+    get_ipython = None
+
+
+@contextmanager
+def suppress_log_scale_warnings():
+    """Context manager to suppress matplotlib log scale warnings when no positive data"""
+    with warnings.catch_warnings():
+        warnings.filterwarnings('ignore', 
+                              message='Data has no positive values, and therefore cannot be log-scaled.',
+                              category=UserWarning)
+        yield
+
+
+class InteractivePlotManager:
+    """
+    Manager for interactive matplotlib figures in Jupyter widgets
+    
+    This class handles the creation and management of persistent matplotlib figures
+    that can be updated in-place without creating new figure windows.
+    
+    Features:
+    - Persistent figure management
+    - Interactive cursor with coordinate display
+    - Zoom and pan capabilities
+    - In-place plot updates
+    - Automatic layout management
+    """
+    
+    def __init__(self):
+        self.figures = {}  # Store persistent figures
+        self.axes = {}     # Store axes references
+        self.cursors = {}  # Store cursor objects
+        self.coord_displays = {}  # Store coordinate display widgets
+        self._setup_interactive_backend()
+    
+    def _setup_interactive_backend(self):
+        """Setup interactive matplotlib backend for Jupyter"""
+        try:
+            # Enable interactive matplotlib widget backend
+            if get_ipython is not None:
+                get_ipython().run_line_magic('matplotlib', 'widget')
+            else:
+                plt.ion()  # Turn on interactive mode
+        except (NameError, AttributeError):
+            # Not in IPython environment, use standard backend
+            plt.ion()  # Turn on interactive mode
+    
+    def create_figure(self, 
+                     figure_id: str, 
+                     figsize: Tuple[float, float] = (12, 6),
+                     plot_sld_profile: bool = False) -> Tuple[plt.Figure, plt.Axes]:
+        """
+        Create or retrieve a persistent figure with interactive features
+        
+        Args:
+            figure_id: Unique identifier for the figure
+            figsize: Figure size as (width, height)
+            plot_sld_profile: Whether to create dual-axis layout for SLD profile
+            
+        Returns:
+            Tuple of (figure, axes) where axes is single axis or tuple of axes
+        """
+        # Close existing figure if it exists
[…]
