+++ pyproject.toml
@@ -36,6 +36,7 @@ dependencies = [
     "click",
     "matplotlib",
     "ipywidgets",
+    "ipympl",
     "huggingface_hub",
     "safetensors",
     "tensorboard"
