+++ reflectorch/extensions/jupyter/widget.py
@@ -1,0 +1,366 @@
+"""
+Reflectorch Jupyter Widget
+
+This module provides a clean, simplified widget interface for reflectometry analysis.
+The widget features a tabbed interface with organized controls and real-time results.
+
+Key Features:
+- Professional tabbed interface
+- Interactive parameter table with live results
+- Comprehensive preprocessing and plotting controls
+- Interactive plotting with coordinate display
+- Clean, modular architecture
+"""
+
+import numpy as np
+from typing import Optional, Dict, Any, Union
+import ipywidgets as widgets
+from IPython.display import display
+
+from .components import (
+    ParameterTable, 
+    PreprocessingControls, 
+    PredictionControls, 
+    PlottingControls, 
+    WidgetSettingsExtractor
+)
+from .interactive_plotting import InteractivePlotManager, plot_reflectivity_interactive
+from ...inference.plotting import plot_reflectivity, print_prediction_results
+
+
+class ReflectorchWidget:
+    """
+    Interactive Jupyter Widget for Reflectometry Analysis
+    
+    A modern, tabbed interface for reflectometry data analysis with Reflectorch models.
+    Features real-time parameter updates, interactive plotting, and comprehensive controls.
+    
+    Attributes:
+        model: The EasyInferenceModel instance
+        prediction_result: Latest prediction results
+        interactive_plots: Whether interactive plotting is enabled
+        
+    Example:
+        ```python
+        from reflectorch.inference import EasyInferenceModel
+        from reflectorch.extensions.jupyter import ReflectorchWidget
+        
+        model = EasyInferenceModel('config.yaml')
+        widget = ReflectorchWidget(model)
+        
+        widget.display(
+            reflectivity_curve=data,
+            q_values=q_values,
+            sigmas=sigmas
+        )
+        
+        # Access results
+        results = widget.prediction_result
+        ```
+    """
+    
+    def __init__(self, model, interactive_plots: bool = True):
+        """
+        Initialize the Reflectorch widget
+        
+        Args:
+            model: EasyInferenceModel instance for making predictions
+            interactive_plots: Enable interactive plotting features (default: True)
+        """
+        self.model = model
+        self.interactive_plots = interactive_plots
+        self.prediction_result = None
+        self.plot_manager = InteractivePlotManager() if interactive_plots else None
+        
+        # Widget components (initialized when display is called)
+        self.parameter_table = None
+        self.preprocessing_controls = None
+        self.prediction_controls = None
+        self.plotting_controls = None
+        
+        self._validate_model()
+    
+    def _validate_model(self):
+        """Validate that the model has required attributes"""
+        required_attrs = ['trainer', 'preprocess_and_predict']
+        for attr in required_attrs:
+            if not hasattr(self.model, attr):
+                raise ValueError(f"Model must have '{attr}' attribute")
+    
+    def display(self, 
+                reflectivity_curve: np.ndarray,
+                q_values: np.ndarray,
+                sigmas: Optional[np.ndarray] = None,
+                q_resolution: Optional[Union[float, np.ndarray]] = None,
+                initial_prior_bounds: Optional[np.ndarray] = None,
+                ambient_sld: Optional[float] = None):
+        """
+        Display the widget interface
+        
+        Args:
+            reflectivity_curve: Experimental reflectivity data
+            q_values: Momentum transfer values (required)
+            sigmas: Experimental uncertainties (optional)
+            q_resolution: Q-resolution, float or array (optional)
+            initial_prior_bounds: Initial bounds for priors, shape (n_params, 2)
+            ambient_sld: Ambient SLD value (optional)
+        """
+        if q_values is None:
[…]

[…]
