### 2025-09-10 b17b10d
create initial plot
