+++ reflectorch/extensions/jupyter/plotly_plot_manager.py
@@ -1,0 +1,329 @@
+import numpy as np
+
+import plotly.graph_objects as go
+
+
+class PlotlyPlotManager:
+    """
+    Manager for Plotly figures in Jupyter widgets
+    """
+    
+    def __init__(self, verbose: bool = False):
+        self.figures = {}  # Store persistent figures
+        self.widgets = {}  # Store plotly widgets
+        self.verbose = verbose
+            
+    def create_reflectivity_figure(self, 
+                                  figure_id: str,
+                                  width: int = 600,
+                                  height: int = 300):
+        """Create a reflectivity-only figure widget"""
+        fig = go.Figure()
+        
+        fig.update_layout(
+            width=width,
+            height=height,
+            showlegend=True,
+            hovermode='closest',
+            template='plotly_white',
+            margin=dict(l=60, r=20, t=60, b=60),
+            legend=dict(
+                orientation="h",
+                yanchor="bottom",
+                y=1.02,
+                xanchor="left",
+                x=0
+            )
+        )
+        
+        # Create Plotly widget
+        plotly_widget = go.FigureWidget(fig)
+        
+        # Store references
+        self.figures[figure_id] = fig
+        self.widgets[figure_id] = plotly_widget
+        
+        return plotly_widget
+    
+    def create_sld_figure(self, 
+                         figure_id: str,
+                         width: int = 600,
+                         height: int = 250):
+        """Create an SLD-only figure widget"""
+        fig = go.Figure()
+        
+        fig.update_layout(
+            width=width,
+            height=height,
+            showlegend=True,
+            hovermode='closest',
+            template='plotly_white',
+            margin=dict(l=60, r=20, t=60, b=60),
+            legend=dict(
+                orientation="h",
+                yanchor="bottom",
+                y=1.02,
+                xanchor="left",
+                x=0
+            )
+        )
+        
+        # Create Plotly widget
+        plotly_widget = go.FigureWidget(fig)
+        
+        # Store references
+        self.figures[figure_id] = fig
+        self.widgets[figure_id] = plotly_widget
+        
+        return plotly_widget
+    
+    def _setup_figure_hover(self, figure_id: str, plotly_widget):
+        """Setup hover functionality for the figure (no coordinate display)"""
+        # Just ensure data is mutable for future trace additions
+        plotly_widget.data = list(plotly_widget.data)
+    
+    def get_figure(self, figure_id: str):
+        """Get existing figure widget"""
+        if figure_id not in self.widgets:
+            raise ValueError(f"Figure '{figure_id}' not found. Create it first with create_figure().")
+        return self.widgets[figure_id]
+    
+    def get_widget(self, figure_id: str):
+        """Get the plotly widget for display"""
+        if figure_id not in self.widgets:
+            raise ValueError(f"Widget for figure '{figure_id}' not found.")
+        return self.widgets[figure_id]
+    
+    
+    def clear_figure(self, figure_id: str):
+        """Clear all traces from the figure"""
+        if figure_id in self.widgets:
+            widget = self.widgets[figure_id]
+            # Use Plotly's proper method to clear traces
+            with widget.batch_update():
+                widget.data = []
+    
+    def close_figure(self, figure_id: str):
+        """Close and cleanup a figure"""
+        if figure_id in self.figures:
+            del self.figures[figure_id]
[…]
