### 2025-09-21 136da54
extended support for huggingface repositories with new layout
