### 2025-09-25 ae5af0e
Update inference_model.py: added chi^2 and reduced chi^2 to polishing step
- If experimental `error_bars` (sigmas) are provided, it computes the weighted chi-squared and the reduced chi-squared (similar to in Refnx) .
- If no error bars are available, it calculates the unweighted sum of squared errors.
