### 2025-09-29 b25fdf2
Update scipy_fitter.py
added parallel  polishing/fitting using joblib
