### 2025-09-29 19ec9d1
Update scipy_fitter.py
added docstring
