### 2025-09-29 42e3938
Update scipy_fitter.py
