+++ reflectorch/models/networks/__init__.py
@@ -4,6 +4,7 @@ from reflectorch.models.networks.mlp_networks import (
     NetworkWithPriorsFnoEmb,
 )
 from reflectorch.models.networks.residual_net import ResidualMLP
+from reflectorch.models.networks.nf_network import NFNetwork
 
 
 __all__ = [
@@ -11,4 +12,5 @@ __all__ = [
     "NetworkWithPriors",
     "NetworkWithPriorsConvEmb",
     "NetworkWithPriorsFnoEmb",
+    "NFNetwork"
 ]
