### 2025-10-09 7674eb8
added method to compute a batch of reflectivity curves in smaller sub-batches
