### 2025-10-09 80886ca
added support for inference with NF models
