+++ reflectorch/inference/inference_model.py
@@ -21,7 +21,7 @@ from reflectorch.runs.utils import (
     get_trainer_by_name, train_from_config
 )
 from reflectorch.runs.config import load_config
-from reflectorch.ml.trainers import PointEstimatorTrainer
+from reflectorch.ml.trainers import PointEstimatorTrainer, NFlowTrainer
 from reflectorch.data_generation.likelihoods import LogLikelihood
 
 from reflectorch.inference.preprocess_exp import StandardPreprocessing
@@ -330,6 +330,9 @@ class EasyInferenceModel(object):
             dict: dictionary containing the predictions
         """
 
+        if not isinstance(self.trainer, PointEstimatorTrainer):
+            raise RuntimeError("`predict()` is only supported for models trained with `PointEstimatorTrainer`.")
+
         scaled_curve = self._scale_curve(reflectivity_curve)
         prior_bounds = np.array(prior_bounds)
 
@@ -436,6 +439,321 @@ class EasyInferenceModel(object):
                 prediction_dict["param_names"].append("max_d_change")
 
         if ambient_sld and not supress_sld_amb_back_shift: #Note: the SLD shift will only be reflected in predicted_params_array but not in predicted_params_object; supress_sld_amb_back_shift is required for the 'preprocess_and_predict' method
+            self._restore_slds_after_ambient_shift(prediction_dict, sld_indices, ambient_sld)   
+
+        return prediction_dict
+    
+    def preprocess_and_sample(self, 
+                            reflectivity_curve: Union[np.ndarray, Tensor], 
+                            q_values: Union[np.ndarray, Tensor] = None, 
+                            prior_bounds: Union[np.ndarray, List[Tuple]] = None, 
+                            sigmas: Union[np.ndarray, Tensor] = None,
+                            q_resolution: float = None,
+                            ambient_sld: float = None,
+                            num_samples: int = 1000,
+                            sampling_batch_size: int = None,
+                            sld_profile_padding_left: float = 0.2,
+                            sld_profile_padding_right: float = 1.1,
+                            kwargs_param_labels: dict = {},
+                            calc_sampled_curves: bool = False,
+                            maximum_sim_batch_size: int = None,
+                            calc_sampled_sld_profiles: bool = False,
+                            calc_log_likelihoods: bool = False,
+                            clip_prediction: bool = False,
+                            enable_importance_sampling: bool = False,
+                            rel_err_factor: float = 0.2,
+
+                            truncate_index_left: int = None,
+                            truncate_index_right: int = None,
+                            enable_error_bars_filtering: bool = True,
+                            filter_threshold=0.3,
+                            filter_remove_singles=True,
+                            filter_remove_consecutives=True,
+                            filter_consecutive=3,
+                            filter_q_start_trunc=0.1,
+                            ):
+        
+        ## Preprocess the data (filtering, truncation, error bar filtering)
+        (q_values, reflectivity_curve, sigmas, q_resolution, 
+        _, _, _, _) = self._preprocess_input_data(
+            reflectivity_curve=reflectivity_curve,
+            q_values=q_values,
+            sigmas=sigmas,
+            q_resolution=q_resolution,
+            truncate_index_left=truncate_index_left,
+            truncate_index_right=truncate_index_right,
+            enable_error_bars_filtering=enable_error_bars_filtering,
+            filter_threshold=filter_threshold,
+            filter_remove_singles=filter_remove_singles,
+            filter_remove_consecutives=filter_remove_consecutives,
[…]
