### 2025-10-09 5ee021f
added nf trainer
