### 2025-10-09 7a7d756
added missing code snippets
