### 2025-10-13 0afa2c3
Polishing with refl_fit returns variance of fit parameters
