### 2025-10-13 9460492
refnx_fit returns error bars, plotting of the results also shows error if applicable, #33
