### 2025-10-14 2e33613
Added support for layer-wise prior bounds input and  integration with huggingface configuration filtering
