+++ reflectorch/inference/input_interface.py
@@ -1,0 +1,165 @@
+from reflectorch.data_generation.priors.parametric_models import NuisanceParamsWrapper
+from reflectorch.inference import inference_model
+
+
+class Layer():
+    def __init__(self, thickness_bounds, roughness_bounds, sld_bounds, imag_sld_bounds=None):
+        self.thickness_bounds = thickness_bounds
+        self.roughness_bounds = roughness_bounds
+        self.sld_bounds = sld_bounds
+        self.imag_sld_bounds = imag_sld_bounds
+
+class Backing():
+    def __init__(self, roughness_bounds, sld_bounds, imag_sld_bounds=None):
+        self.thickness_bounds = None
+        self.roughness_bounds = roughness_bounds
+        self.sld_bounds = sld_bounds
+        self.imag_sld_bounds = imag_sld_bounds
+
+class Structure():
+    def __init__(self, layers, q_shift_bounds=None, r_scale_bounds=None, log10_background_bounds=None):
+        self.layers=layers
+        self.q_shift_bounds=q_shift_bounds
+        self.r_scale_bounds = r_scale_bounds
+        self.log10_background_bounds = log10_background_bounds
+        self.thicknesses_bounds = []
+        self.roughnesses_bounds = []
+        self.slds_bounds = []
+        self.imag_slds_bounds = []
+
+        for layer in layers:
+            if layer.thickness_bounds is not None:
+                self.thicknesses_bounds.append(layer.thickness_bounds)
+            self.roughnesses_bounds.append(layer.roughness_bounds)
+            self.slds_bounds.append(layer.sld_bounds)
+            if layer.imag_sld_bounds is not None:
+                self.imag_slds_bounds.append(layer.imag_sld_bounds)
+        
+        self.prior_bounds = self.thicknesses_bounds + self.roughnesses_bounds + self.slds_bounds + self.imag_slds_bounds
+
+        if q_shift_bounds is not None:
+            self.prior_bounds += [q_shift_bounds]
+        if r_scale_bounds is not None:
+            self.prior_bounds += [r_scale_bounds]
+        if log10_background_bounds is not None:
+            self.prior_bounds += [log10_background_bounds]
+    
+    def validate_parameters_and_ranges(self, inference_model):
+        if len(self.layers) - 1 != inference_model.trainer.loader.prior_sampler.max_num_layers:
+            raise ValueError(f'Number of layers mismatch: this model expects {inference_model.trainer.loader.prior_sampler.max_num_layers} layers (backing not included) but you provided {len(self.layers) - 1}')
+        
+
+        thickness_ranges = inference_model.trainer.loader.prior_sampler.param_ranges['thicknesses']
+        roughness_ranges = inference_model.trainer.loader.prior_sampler.param_ranges['roughnesses']
+        sld_ranges = inference_model.trainer.loader.prior_sampler.param_ranges['slds']
+        
+        def layer_name(i):
+            if i == inference_model.trainer.loader.prior_sampler.max_num_layers:
+                return 'the backing medium'
+            else:
+                return f'layer {i+1}'
+
+        for i, layer in enumerate(self.layers):
+            if layer.thickness_bounds is not None:
+                if layer.thickness_bounds[0] < thickness_ranges[0] or layer.thickness_bounds[1] > thickness_ranges[1]:
+                    raise ValueError(f"The provided prior bounds for the thickness of layer {i+1} are outside the training range of the network: {thickness_ranges}")
+            if layer.roughness_bounds[0] < roughness_ranges[0] or layer.roughness_bounds[1] > roughness_ranges[1]:
+                raise ValueError(f"The provided prior bounds for the roughness of {layer_name(i)} are outside the training range of the network: {roughness_ranges}")
[…]
