### 2025-10-14 2d56169
widget new model selection
