+++ reflectorch/extensions/jupyter/model_selection.py
@@ -1,0 +1,523 @@
+"""
+Jupyter Widget Model Selection Component for Reflectorch
+
+This module contains the model browsing and selection component for
+reflectometry analysis models from Hugging Face repositories.
+
+Components:
+    - ModelSelection: Model browsing and selection from Hugging Face
+"""
+
+from typing import Optional, Dict, Any, Callable
+import ipywidgets as widgets
+import threading
+from huggingface_hub import HfApi
+
+from reflectorch.extensions.jupyter.custom_select import CustomSelect
+
+class ModelSelection:
+    """Model selection component for Hugging Face models.
+    """
+    
+    def __init__(self, organization_list: tuple[str, ...] = ('reflectorch-ILL', )):
+        """
+        Initialize model selection component
+        
+        Args:
+            organization_list: Tuple of Hugging Face organizations to browse models from
+        """
+        assert len(organization_list) > 0, "At least one organization must be provided"
+        self.organization = organization_list[0]
+        self.organization_list = organization_list
+        self.hf_api = HfApi()
+        self.models_data = []
+        self.selected_config = None
+        self.selected_model = None
+        self._model_cache = {}  # Cache model info by organization
+        self._download_callback = None  # External download handler
+        
+        self.widget = self._create_model_browser()
+        
+    def _create_model_details_template(self) -> str:
+        """Create a comprehensive HTML template for model details"""
+        return """
+        <div style="font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', sans-serif; line-height: 1.5;">
+            <div style="margin-bottom: 12px;">
+                <h4 style="margin: 0 0 8px 0; color: #0d6efd; font-size: 16px;">📋 {modelId}</h4>
+                <div style="font-size: 13px; color: #6c757d;">
+                    <b>Layers:</b> {num_layers} | <b>Parameterization:</b> {parameterization}
+                </div>
+            </div>
+            
+            <div style="margin-bottom: 12px;">
+                <h5 style="margin: 0 0 6px 0; color: #495057; font-size: 14px;">Parameter Ranges</h5>
+                <div style="background: #f8f9fa; border: 1px solid #dee2e6; border-radius: 4px; padding: 8px; font-size: 12px;">
+                    {param_ranges_table}
+                </div>
+            </div>
+            
+            <div style="margin-bottom: 12px;">
+                <h5 style="margin: 0 0 6px 0; color: #495057; font-size: 14px;">Prior Bound Widths</h5>
+                <div style="background: #f8f9fa; border: 1px solid #dee2e6; border-radius: 4px; padding: 8px; font-size: 12px;">
+                    {bound_width_ranges_table}
+                </div>
+            </div>
+            
+            <div style="margin-bottom: 12px;">
+                <h5 style="margin: 0 0 6px 0; color: #495057; font-size: 14px;">Misalignment Support</h5>
+                <div style="background: #f8f9fa; border: 1px solid #dee2e6; border-radius: 4px; padding: 8px; font-size: 12px;">
+                    {misalignment_info}
+                </div>
+            </div>
+            
+            <div>
+                <h5 style="margin: 0 0 6px 0; color: #495057; font-size: 14px;">Additional Info</h5>
+                <div style="background: #f8f9fa; border: 1px solid #dee2e6; border-radius: 4px; padding: 8px; font-size: 12px;">
+                    {additional_metadata}
+                </div>
+            </div>
+        </div>
+        """
+        
+    def _create_model_browser(self) -> widgets.VBox:
+        """Create the model browser widget"""
+        # Title and description
+        title = widgets.HTML(
+            "<h3>🤗 Model Selection</h3>",
+            layout=widgets.Layout(margin='0px 0px 10px 0px')
+        )
[…]
