+++ reflectorch/extensions/jupyter/components.py
@@ -424,3 +424,5 @@ class WidgetSettingsExtractor:
         
         search_widget(container)
         return found_widgets
+
+
