### 2025-10-14 60b539f
model loading works
