### 2025-10-16 72c8bfc
Updated docstrings for the inference methods, added input safeguards, renamed EasyInferenceModel to InferenceModel (created an alias for back-compatibility)
