+++ reflectorch/inference/inference_model.py
@@ -31,7 +31,7 @@ from reflectorch.inference.record_time import print_time
 from reflectorch.inference.plotting import plot_reflectivity, plot_prediction_results, print_prediction_results
 from reflectorch.utils import get_filtering_mask, to_t
 
-class EasyInferenceModel(object):
+class InferenceModel(object):
     """Facilitates the inference process using pretrained models
     
     Args:
@@ -159,11 +159,11 @@ class EasyInferenceModel(object):
             print(f"The following quantities are additional inputs to the network: {inputs_str}.")
 
     def preprocess_and_predict(self, 
-                            reflectivity_curve: Union[np.ndarray, Tensor], 
-                            q_values: Union[np.ndarray, Tensor] = None, 
+                            reflectivity_curve: np.ndarray, 
+                            q_values: np.ndarray = None, 
                             prior_bounds: Union[np.ndarray, List[Tuple]] = None, 
-                            sigmas: Union[np.ndarray, Tensor] = None,
-                            q_resolution: float = None,
+                            sigmas: np.ndarray = None,
+                            q_resolution: Union[float, np.ndarray] = None,
                             ambient_sld: float = None,
                             clip_prediction: bool = True, 
                             polish_prediction: bool = False,
@@ -189,6 +189,37 @@ class EasyInferenceModel(object):
                             filter_consecutive=3,
                             filter_q_start_trunc=0.1,
                             ):
+        """Preprocess experimental data (clean, truncate, filter, interpolate) and run prediction. This wrapper prepares inputs according to the model's Q generator calls `predict(...)` on the interpolated/padded data, and (optionally) performs a polishing step on the original data (pre-interpolation) 
+
+        Args:
+            reflectivity_curve (Union[np.ndarray, Tensor]): 1D array of experimental reflectivity values.
+            q_values (Union[np.ndarray, Tensor]): 1D array of momentum transfer values for the reflectivity curve (in units of inverse angstroms).
+            prior_bounds (Union[np.ndarray, List[Tuple]]): Prior bounds for all parameters, shape ``(num_params, 2)`` as ``[(min, max), …]``.
+            sigmas (Union[np.ndarray, Tensor], optional): 1D array of experimental uncertainties (same length as `reflectivity_curve`). Used for error-bar filtering (if enabled) and for polishing (if requested).
+            q_resolution (Union[float, np.ndarray], optional): The q resolution for neutron reflectometry models. Can be either a float (dq/q) for linear resolution smearing (e.g. 0.05 meaning 5% reolution smearing) or an array of dq values for pointwise resolution smearing. 
+            ambient_sld (float, optional): The SLD of the fronting (i.e. ambient) medium for structure with fronting medium different than air.
+            clip_prediction (bool, optional): If ``True``, the values of the predicted parameters are clipped to not be outside the interval set by the prior bounds. Defaults to True.
+            polish_prediction (bool, optional): If ``True``, the neural network predictions are further polished using a simple least mean squares (LMS) fit. Defaults to False.
+            polishing_method (str): {'trf', 'dogbox', 'lm'} SciPy least-squares method used for polishing.
+            use_sigmas_for_polishing (bool): If ``True``, weigh residuals by `sigmas` during polishing.
+            polishing_max_nfev (int, optional): Maximum number of function evaluations for the SciPy optimizer.
+            fit_growth (bool, optional): (Deprecated) If ``True``, an additional parameters is introduced during the LMS polishing to account for the change in the thickness of the upper layer during the in-situ measurement of the reflectivity curve (a linear growth is assumed). Defaults to False.
[…]
