+++ reflectorch/extensions/jupyter/components.py
@@ -9,6 +9,7 @@ Components:
     - PreprocessingControls: Data preprocessing options
     - PredictionControls: Prediction and computation settings
     - PlottingControls: Plotting and visualization options
+    - AdditionalParametersControls: Controls for additional parameters like Q resolution
 """
 
 import numpy as np
@@ -29,7 +30,8 @@ class ParameterTable:
     
     def __init__(self, param_labels: List[str], min_bounds: np.ndarray, 
                  max_bounds: np.ndarray, max_deltas: np.ndarray, 
-                 initial_bounds: Optional[np.ndarray] = None):
+                 initial_bounds: Optional[np.ndarray] = None,
+                 additional_params_controls: Optional['AdditionalParametersControls'] = None):
         """
         Initialize parameter table
         
@@ -39,6 +41,7 @@ class ParameterTable:
             max_bounds: Maximum values for each parameter  
             max_deltas: Maximum allowed range for each parameter
             initial_bounds: Initial bounds, shape (n_params, 2)
+            additional_params_controls: Optional additional parameters component
         """
         self.param_labels = param_labels
         self.min_bounds = min_bounds
@@ -46,6 +49,7 @@ class ParameterTable:
         self.max_deltas = max_deltas
         self.sliders = []
         self.result_displays = {}
+        self.additional_params_controls = additional_params_controls
         
         self.widget = self._create_table(initial_bounds)
     
@@ -115,13 +119,26 @@ class ParameterTable:
             
             parameter_rows.append(row)
         
-        # Create the complete table
-        return widgets.VBox([
+        # Create the main parameter table
+        main_table = widgets.VBox([
             widgets.HTML("<h4>Parameter Configuration</h4>"),
             header,
             widgets.HTML("<hr style='margin: 5px 0px;'>"),
             *parameter_rows
         ])
+        
+        # Create the complete widget with optional additional parameters
+        table_components = [main_table]
+        
+        # Add additional parameters section if available
+        if (self.additional_params_controls is not None and 
+            self.additional_params_controls.additional_sliders):
+            table_components.extend([
+                widgets.HTML("<br>"),
+                self.additional_params_controls.widget
+            ])
+        
+        return widgets.VBox(table_components)
     
     def _add_slider_validation(self, slider: widgets.FloatRangeSlider, max_width: float):
         """Add validation to constrain slider range"""
@@ -257,6 +274,68 @@ class PredictionControls:
         ])
 
 
+class AdditionalParametersControls:
+    """
+    Controls for additional parameters that are not part of prior bounds
+    
+    These are fixed input parameters like Q resolution that are passed
+    separately to the inference model.
+    """
+    
+    def __init__(self, inference_model):
+        """
+        Initialize additional parameters controls
+        
+        Args:
+            inference_model: The inference model to check for additional parameters
+        """
+        self.inference_model = inference_model
+        self.additional_sliders = {}
+        self.widget = self._create_controls()
+    
+    def _create_controls(self) -> widgets.VBox:
+        """Create additional parameters controls widget"""
+        controls = [widgets.HTML("<h4>Additional Parameters</h4>")]
+        
+        # Check if model has smearing (Q resolution)
+        if (hasattr(self.inference_model, 'trainer') and 
+            hasattr(self.inference_model.trainer, 'loader') and
+            hasattr(self.inference_model.trainer.loader, 'smearing') and
+            self.inference_model.trainer.loader.smearing is not None):
+            
+            q_res_min = float(self.inference_model.trainer.loader.smearing.sigma_min)
+            q_res_max = float(self.inference_model.trainer.loader.smearing.sigma_max)
+            
[…]
