### 2025-10-16 939d874
better widget api
