+++ reflectorch/extensions/jupyter/widget.py
@@ -56,16 +56,41 @@ class ReflectorchPlotlyWidget:
         ```
     """
     
-    def __init__(self, model):
+    def __init__(self,                
+                reflectivity_curve: np.ndarray,
+                q_values: np.ndarray,
+                sigmas: Optional[np.ndarray] = None,
+                q_resolution: Optional[Union[float, np.ndarray]] = None,
+                initial_prior_bounds: Optional[np.ndarray] = None,
+                ambient_sld: Optional[float] = None,
+                model: Optional["InferenceModel"] = None,
+):
         """
         Initialize the Reflectorch Plotly widget
         
         Args:
-            model: InferenceModel instance for making predictions
+            reflectivity_curve: Experimental reflectivity data
+            q_values: Momentum transfer values
+            sigmas: Experimental uncertainties (optional)
+            q_resolution: Q-resolution, float or array (optional)
+            initial_prior_bounds: Initial bounds for priors, shape (n_params, 2)
+            ambient_sld: Ambient SLD value (optional)
+            model: InferenceModel instance for making predictions (optional)
         """
         self.model = model
         self.prediction_result = None
         self.plot_manager = PlotlyPlotManager()
+
+        # Store data for prediction
+        self._data = {
+            'reflectivity_curve': reflectivity_curve,
+            'q_values': q_values,
+            'sigmas': sigmas,
+            'q_resolution': q_resolution,
+            'ambient_sld': ambient_sld
+        }
+
+        self.initial_prior_bounds = initial_prior_bounds
         
         # Widget components (initialized when display is called)
         self.parameter_table = None
@@ -76,7 +101,41 @@ class ReflectorchPlotlyWidget:
         self.model_selection = None
         self.tabs_widget = None  # Store reference to tabs for updating
         
-        self._validate_model()
+        if self.model is not None:
+            self._validate_model()
+    
+    def _create_parameter_components(self, initial_prior_bounds=None):
+        """
+        Create parameter table and additional parameters controls based on current model
+        
+        Args:
+            initial_prior_bounds: Initial bounds for priors, shape (n_params, 2)
+            
+        Returns:
+            Tuple of (parameter_table, additional_params_controls)
+        """
+        # Create additional params controls first so we can pass it to parameter table
+        additional_params_controls = AdditionalParametersControls(self.model)
+        
+        # Get model parameters info if model is available
+        if self.model is not None:
+            param_labels = self.model.trainer.loader.prior_sampler.param_model.get_param_labels()
+            min_bounds = self.model.trainer.loader.prior_sampler.min_bounds.cpu().numpy().flatten()
+            max_bounds = self.model.trainer.loader.prior_sampler.max_bounds.cpu().numpy().flatten()
+            max_deltas = self.model.trainer.loader.prior_sampler.max_delta.cpu().numpy().flatten()
+        else:
+            # Default empty parameters when no model is loaded
+            param_labels = []
+            min_bounds = np.array([])
+            max_bounds = np.array([])
+            max_deltas = np.array([])
+        
+        parameter_table = ParameterTable(
+            param_labels, min_bounds, max_bounds, max_deltas, initial_prior_bounds,
+            additional_params_controls=additional_params_controls
+        )
+        
+        return parameter_table, additional_params_controls
     
     def _validate_model(self):
         """Validate that the model has required attributes"""
@@ -86,12 +145,6 @@ class ReflectorchPlotlyWidget:
                 raise ValueError(f"Model must have '{attr}' attribute")
     
     def display(self, 
-                reflectivity_curve: np.ndarray,
-                q_values: np.ndarray,
-                sigmas: Optional[np.ndarray] = None,
-                q_resolution: Optional[Union[float, np.ndarray]] = None,
-                initial_prior_bounds: Optional[np.ndarray] = None,
[…]
