+++ reflectorch/extensions/jupyter/components.py
@@ -31,7 +31,8 @@ class ParameterTable:
     def __init__(self, param_labels: List[str], min_bounds: np.ndarray, 
                  max_bounds: np.ndarray, max_deltas: np.ndarray, 
                  initial_bounds: Optional[np.ndarray] = None,
-                 additional_params_controls: Optional['AdditionalParametersControls'] = None):
+                 additional_params_controls: Optional['AdditionalParametersControls'] = None,
+                 predict_button: Optional[widgets.Button] = None):
         """
         Initialize parameter table
         
@@ -42,6 +43,7 @@ class ParameterTable:
             max_deltas: Maximum allowed range for each parameter
             initial_bounds: Initial bounds, shape (n_params, 2)
             additional_params_controls: Optional additional parameters component
+            predict_button: Optional predict button to include in the table header
         """
         self.param_labels = param_labels
         self.min_bounds = min_bounds
@@ -50,6 +52,7 @@ class ParameterTable:
         self.sliders = []
         self.result_displays = {}
         self.additional_params_controls = additional_params_controls
+        self.predict_button = predict_button
         
         self.widget = self._create_table(initial_bounds)
     
@@ -126,10 +129,21 @@ class ParameterTable:
             ], layout=widgets.Layout(margin='2px 0px'))
             
             parameter_rows.append(row)
+
+        self.param_title = widgets.HTML("<h4>Parameter Configuration</h4>")
+        
+        # Create title row with optional predict button
+        if self.predict_button is not None:
+            title_row = widgets.HBox([
+                self.param_title,
+                self.predict_button
+            ], layout=widgets.Layout(justify_content='space-between', align_items='center'))
+        else:
+            title_row = self.param_title
         
         # Create the main parameter table
         main_table = widgets.VBox([
-            widgets.HTML("<h4>Parameter Configuration</h4>"),
+            title_row,
             header,
             widgets.HTML("<hr style='margin: 5px 0px;'>"),
             *parameter_rows
@@ -213,16 +227,14 @@ class PreprocessingControls:
             # Truncation section
             widgets.HTML("<h5>Data Truncation</h5>"),
             widgets.HTML("<i>Specify which data points to include in the analysis</i>"),
-            widgets.HBox([
+            widgets.VBox([
                 widgets.IntSlider(
                     description='Left index:', min=0, max=max(0, self.n_datapoints-1), 
-                    step=1, value=0, style={'description_width': '100px'},
-                    layout=widgets.Layout(width='300px')
+                    step=1, value=0,
                 ),
                 widgets.IntSlider(
                     description='Right index:', min=1, max=self.n_datapoints, 
-                    step=1, value=self.n_datapoints, style={'description_width': '100px'},
-                    layout=widgets.Layout(width='300px')
+                    step=1, value=self.n_datapoints,
                 )
             ]),
             
@@ -231,23 +243,20 @@ class PreprocessingControls:
             # Error bar filtering section  
             widgets.HTML("<h5>Error Bar Filtering</h5>"),
             widgets.HTML("<i>Filter out unreliable data points based on error bars</i>"),
-            widgets.HBox([
-                widgets.Checkbox(description='Enable filtering', value=True, style={'description_width': '120px'}),
-                widgets.Checkbox(description='Remove singles', value=True, style={'description_width': '120px'}),
-                widgets.Checkbox(description='Remove consecutives', value=True, style={'description_width': '140px'})
+            widgets.VBox([
+                widgets.Checkbox(description='Enable filtering', value=True),
+                widgets.Checkbox(description='Remove singles', value=True),
+                widgets.Checkbox(description='Remove consecutives', value=True)
             ]),
[…]
