### 2025-10-16 1c0d057
log widget
