+++ reflectorch/extensions/jupyter/widget.py
@@ -3,7 +3,10 @@ Reflectorch Jupyter Widget
 """
 
 import numpy as np
-from typing import Optional, Union
+from typing import Optional, Union, TYPE_CHECKING
+
+if TYPE_CHECKING:
+    from reflectorch.inference.inference_model import InferenceModel
 import ipywidgets as widgets
 from IPython.display import display
 
@@ -21,6 +24,7 @@ from reflectorch.extensions.jupyter.components import (
     WidgetSettingsExtractor,
 )
 from reflectorch.extensions.jupyter.model_selection import ModelSelection
+from reflectorch.extensions.jupyter.log_widget import LogWidget
 
 
 class ReflectorchPlotlyWidget:
@@ -101,6 +105,7 @@ class ReflectorchPlotlyWidget:
         self.model_selection = None
         self.tabs_widget = None  # Store reference to tabs for updating
         self.predict_button = None  # Store reference to predict button
+        self.log_widget = None  # Store reference to log widget
         
         if self.model is not None:
             self._validate_model()
@@ -147,6 +152,7 @@ class ReflectorchPlotlyWidget:
             if not hasattr(self.model, attr):
                 raise ValueError(f"Model must have '{attr}' attribute")
     
+    
     def display(self, 
                 controls_width: int = 700,
                 plot_width: int = 400,
@@ -186,6 +192,9 @@ class ReflectorchPlotlyWidget:
         self.plotting_controls = PlottingControls()
         self.model_selection = ModelSelection()
         
+        # Create log widget
+        self.log_widget = LogWidget()
+        
         # Create tabbed interface
         # if model is not provided, then Models tabs goes first.
         tabs = widgets.Tab()
@@ -212,9 +221,6 @@ class ReflectorchPlotlyWidget:
         # Store reference to tabs for later updates
         self.tabs_widget = tabs
         
-        # Output areas
-        output = widgets.Output()
-        
         # Create plot containers (initially empty)
         reflectivity_plot_container = widgets.VBox([])
         sld_plot_container = widgets.VBox([])
@@ -230,14 +236,19 @@ class ReflectorchPlotlyWidget:
         
         controls_area = widgets.VBox([
             header,
-            tabs,
-            output
+            tabs
         ], layout=widgets.Layout(width=f'{controls_width}px'))
         
         # Horizontal layout: controls on left, plots on right
-        main_layout = widgets.HBox([
+        main_content = widgets.HBox([
             controls_area,
             plot_area
+        ])
+        
+        # Complete layout with log at the bottom
+        main_layout = widgets.VBox([
+            main_content,
+            self.log_widget.widget
         ])
         
         # Add border around the entire widget
@@ -251,10 +262,10 @@ class ReflectorchPlotlyWidget:
         display(container)
         
         # Setup event handlers
-        self._setup_event_handlers(self.predict_button, output, reflectivity_plot_container, sld_plot_container, container)
+        self._setup_event_handlers(self.predict_button, reflectivity_plot_container, sld_plot_container, container)
         
         # Setup model selection integration
-        self._setup_model_selection_integration(output)
+        self._setup_model_selection_integration()
         
         # Setup truncation synchronization
         self._setup_truncation_sync()
@@ -282,7 +293,10 @@ class ReflectorchPlotlyWidget:
             self._plot_initial_data(settings, reflectivity_container, sld_container, plot_width, plot_height)
             
         except Exception as e:
-            print(f"⚠️  Could not create initial plots: {str(e)}")
+            if self.log_widget:
+                self.log_widget.log(f"⚠️  Could not create initial plots: {str(e)}")
+            else:
+                print(f"⚠️  Could not create initial plots: {str(e)}")
     
[…]
