+++ reflectorch/inference/inference_model.py
@@ -1,7 +1,4 @@
-import asyncio
-import logging
 from pathlib import Path
-import time
 
 import numpy as np
 import torch
@@ -9,24 +6,20 @@ from torch import Tensor
 from typing import List, Tuple, Union
 from huggingface_hub import hf_hub_download
 
-from reflectorch.data_generation.priors import Params, BasicParams, ExpUniformSubPriorSampler, UniformSubPriorParams
+from reflectorch.data_generation.priors import BasicParams
 from reflectorch.data_generation.priors.parametric_models import NuisanceParamsWrapper
 from reflectorch.data_generation.q_generator import ConstantQ, VariableQ, MaskedVariableQ
-from reflectorch.data_generation.utils import get_density_profiles, get_param_labels
+from reflectorch.data_generation.utils import get_density_profiles
 from reflectorch.inference.preprocess_exp.interpolation import interp_reflectivity
-from reflectorch.paths import CONFIG_DIR, ROOT_DIR, SAVED_MODELS_DIR
+from reflectorch.paths import CONFIG_DIR, SAVED_MODELS_DIR
 from reflectorch.runs.utils import (
-    get_trainer_by_name, train_from_config
+    get_trainer_by_name
 )
-from reflectorch.runs.config import load_config
 from reflectorch.ml.trainers import PointEstimatorTrainer
 from reflectorch.data_generation.likelihoods import LogLikelihood
 
-from reflectorch.inference.preprocess_exp import StandardPreprocessing
-from reflectorch.inference.scipy_fitter import standard_refl_fit, refl_fit, get_fit_with_growth
-from reflectorch.inference.sampler_solution import simple_sampler_solution, get_best_mse_param
-from reflectorch.inference.record_time import print_time
-from reflectorch.inference.plotting import plot_reflectivity, plot_prediction_results, print_prediction_results
+from reflectorch.inference.scipy_fitter import refl_fit, get_fit_with_growth
+from reflectorch.inference.sampler_solution import get_best_mse_param
 from reflectorch.utils import get_filtering_mask, to_t
 
 class InferenceModel(object):
