### 2025-10-16 54e8dc2
better table ui and error bars
