### 2025-10-17 c260f69
Update curve fitting parameters in scipy_fitter.py
Replaced 'polishing_max_nfev' with 'max_steps' for curve fitting.
