### 2025-10-17 ef07f3a
Add 'anywidget' to project dependencies
