### 2025-10-18 7e3264f
renamed polishing_nfev  argument to polishing_max_steps
