+++ reflectorch/inference/scipy_fitter.py
@@ -64,7 +64,7 @@ def refl_fit(
         error_bars: np.ndarray = None,
         scale_curve_func=np.log10,
         method: str = 'trf', #'lm', 'trf'
-        polishing_max_nfev: int = None,
+        polishing_max_steps: int = None,
         reflectivity_kwargs: dict = None,
         **kwargs
 ):
@@ -97,6 +97,12 @@ def refl_fit(
     else:
         scaled_error_bars = None  
 
+    if polishing_max_steps is not None:
+        if method == 'lm':
+            kwargs['maxfev'] = polishing_max_steps
+        else:
+            kwargs['max_nfev'] = polishing_max_steps
+
     res = curve_fit(
         f=get_scaled_curve_func(
             scale_curve_func=scale_curve_func,
@@ -109,7 +115,6 @@ def refl_fit(
         sigma=scaled_error_bars,
         absolute_sigma=True,
         method=method,
-        max_nfev=polishing_max_nfev,
         **kwargs
     )
 
