### 2025-10-19 92dd9f2
added docstrings
