+++ reflectorch/inference/input_interface.py
@@ -1,6 +1,16 @@
 from reflectorch.data_generation.priors.parametric_models import NuisanceParamsWrapper
 
 class Layer():
+    """Defines a single slab layer with prior bounds for thickness, roughness and SLD.
+
+    The bounds can be given for both real and imaginary parts of the SLD (the latter only if the model supports absorption).
+
+    Args:
+        thickness_bounds (Tuple[float, float]): Minimum and maximum thickness of the layer (in Å).
+        roughness_bounds (Tuple[float, float]): Minimum and maximum interfacial roughness at the top of this layer (in Å).
+        sld_bounds (Tuple[float, float]): Minimum and maximum real SLD of this layer (in 10⁻⁶ Å⁻²).
+        imag_sld_bounds (Tuple[float, float], optional): Minimum and maximum imaginary SLD (in 10⁻⁶ Å⁻²) of this layer. Defaults to None.
+    """
     def __init__(self, thickness_bounds, roughness_bounds, sld_bounds, imag_sld_bounds=None):
         self.thickness_bounds = thickness_bounds
         self.roughness_bounds = roughness_bounds
@@ -8,6 +18,16 @@ class Layer():
         self.imag_sld_bounds = imag_sld_bounds
 
 class Backing():
+    """Defines the backing medium (substrate) for the multilayer structure.
+
+    The backing is assumed to be semi-infinite and has no thickness parameter. 
+    This class ensures compatibility with the layer-based structure definition.
+
+    Args:
+        roughness_bounds (Tuple[float, float]): Minimum and maximum interfacial roughness at the top of the backing medium (in Å).
+        sld_bounds (Tuple[float, float]): Minimum and maximum real SLD of the backing medium (in 10⁻⁶ Å⁻²).
+        imag_sld_bounds (Tuple[float, float], optional): Minimum and maximum imaginary SLD (in 10⁻⁶ Å⁻²) of the backing. Defaults to None.
+    """
     def __init__(self, roughness_bounds, sld_bounds, imag_sld_bounds=None):
         self.thickness_bounds = None
         self.roughness_bounds = roughness_bounds
@@ -15,6 +35,33 @@ class Backing():
         self.imag_sld_bounds = imag_sld_bounds
 
 class Structure():
+    """Defines a multilayer structure and its parameter bounds in a layer-wise manner.
+
+    This class allows the user to define the prior bounds for the full structure (film + backing) in a layer-wise format. It automatically constructs the
+    flattened list of parameter bounds compatible with the inference model’s expected input format.
+
+    Args:
+        layers (List[Union[Layer, Backing]]): Ordered list of layers defining the structure, from the ambient side to the backing. The last element
+            must be a :class:`Backing` instance. Note that the fronting medium (ambient) is not part of this list (since it is not a predicted parameter), 
+            and is treated by default as being 0 (air). For different fronting media one can use the ``ambient_sld`` argument of the prediction method.
+        q_shift_bounds (Tuple[float, float], optional): Bounds for the global ``q_shift`` nuisance parameter. Defaults to None.
+        r_scale_bounds (Tuple[float, float], optional): Bounds for the global reflectivity scale factor ``r_scale``. Defaults to None.
+        log10_background_bounds (Tuple[float, float], optional): Bounds for the background term expressed as log10(background). Defaults to None.
+
+    Attributes:
+        thicknesses_bounds (List[Tuple[float, float]]): Bounds for all thicknesses (excluding backing).
+        roughnesses_bounds (List[Tuple[float, float]]): Bounds for all roughnesses (including backing).
+        slds_bounds (List[Tuple[float, float]]): Bounds for all real SLDs (including backing).
+        imag_slds_bounds (List[Tuple[float, float]]): Bounds for all imaginary SLDs (if provided).
[…]
