### 2025-10-19 f60db0c
Fix UnboundLocalError in _polish_prediction by initializing polished_params_err
