### 2025-10-19 3cede66
remove threading
