### 2025-10-20 32968d3
disable progress bar in huggingface
