### 2025-10-21 30090b7
better ui and error bars included
