+++ reflectorch/extensions/jupyter/components.py
@@ -49,7 +49,9 @@ class ParameterTable:
         self.min_bounds = min_bounds
         self.max_bounds = max_bounds
         self.max_deltas = max_deltas
-        self.sliders = []
+        self.sliders = []  # Store range sliders
+        self.min_inputs = []  # Store min bound inputs
+        self.max_inputs = []  # Store max bound inputs
         self.result_displays = {}
         self.additional_params_controls = additional_params_controls
         self.predict_button = predict_button
@@ -60,9 +62,12 @@ class ParameterTable:
         """Create the parameter table widget"""
         init_pb = np.array(initial_bounds) if initial_bounds is not None else None
         
-        # Add custom CSS for slider styling
-        slider_style = widgets.HTML("""
+        # Add custom CSS for styling
+        custom_style = widgets.HTML("""
         <style>
+            .widget-float-text input {
+                text-align: center;
+            }
             .widget-inline-hbox .widget-readout {
                 margin-left: 20px !important;
             }
@@ -75,10 +80,12 @@ class ParameterTable:
         # Create header row
         header = widgets.HBox([
             widgets.HTML("<b>Parameter</b>", layout=widgets.Layout(width='150px')),
-            widgets.HTML("<b>Prior Bounds</b>", layout=widgets.Layout(width='400px')),
-            widgets.HTML("<b>Predicted</b>", layout=widgets.Layout(width='120px', margin='0px 0px 0px 20px')),
-            widgets.HTML("<b>Polished</b>", layout=widgets.Layout(width='120px')),
-            widgets.HTML("<b>Uncertainty</b>", layout=widgets.Layout(width='120px'))
+            widgets.HTML("<b>Prior Bounds</b>", layout=widgets.Layout(width='300px')),
+            widgets.HTML("<b></b>", layout=widgets.Layout(width='50px')),
+            widgets.HTML("<b></b>", layout=widgets.Layout(width='50px')),
+            widgets.HTML("<b>Predicted</b>", layout=widgets.Layout(width='100px')),
+            widgets.HTML("<b>Polished</b>", layout=widgets.Layout(width='100px')),
+            widgets.HTML("<b>Uncertainty</b>", layout=widgets.Layout(width='100px'))
         ], layout=widgets.Layout(margin='5px 0px', align_items='center'))
         
         # Create parameter rows
@@ -102,29 +109,46 @@ class ParameterTable:
                 layout=widgets.Layout(width='150px', display='flex', align_items='center', justify_content='flex-start')
             )
             
-            # Prior bounds slider - use container to control spacing
+            # Range slider - updates continuously
             slider = widgets.FloatRangeSlider(
                 value=[init_min, init_max],
                 min=float(self.min_bounds[i]),
                 max=float(self.max_bounds[i]),
                 step=0.01,
-                layout=widgets.Layout(width='400px'),
-                readout_format='.2f',
+                layout=widgets.Layout(width='180px'),
+                readout=False,  # Hide readout since we have input boxes
+                continuous_update=True,  # Update instantly while dragging
                 style={'description_width': '0px', 'handle_color': '#2196F3'}
+            )
+            
+            # Min bound input - updates continuously
+            min_input = widgets.FloatText(
+                value=round(init_min, 2),
+                step=0.01,
+                layout=widgets.Layout(width='65px'),
+                continuous_update=True  # Update on every keystroke
+            )
+            
+            # Max bound input - updates continuously
+            max_input = widgets.FloatText(
+                value=round(init_max, 2),
+                step=0.01,
+                layout=widgets.Layout(width='65px'),
+                continuous_update=True  # Update on every keystroke
             )
             
             # Result displays
             predicted_display = widgets.HTML(
                 value="<i>-</i>",
[…]
