### 2025-11-03 #8169 @crvernon
👋  @valentinsingularity - I just need you to make the following modifications to your paper which is generally in great shape:

- [ ] LINE 39-40:  You currently state, "...as described in (Munteanu et al., 2024), the user..." but the in-text reference in this case should be formatted as the following:  "...as described in Munteanu et al. (2024), the user..."  You can do this by simply removing the brackets around this reference in your paper.
- [ ] LINE 63:  "Huggingface" in this context should be written as "Hugging Face"
- [ ] LINE 71:  "Tensorflow" should be written "TensorFlow"
- [ ] LINE 73:  "...the publication (Schumi-Mareček et al., 2024)." should be formatted as "the publication by Schumi-Mareček et al. (2024)." where "by" is also added in.

- [ ]  Conduct a GitHub release of the current reviewed version of the software and archive the reviewed software in Zenodo or a similar service (e.g., figshare, an institutional repository).  Please ensure that the software archive uses the same license as the license you have posted on code repository.

- [ ]  Check the archival deposit (e.g., in Zenodo) to ensure it has the correct metadata. This includes the title (should match the paper title exactly) and author list (should match the people and the order in which they are listed on your paper exactly).

- [ ]  Please respond with the DOI of the archived version here along with the version number released.
